% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.zstat.in.genesOut.r
\name{adjust.zstat.in.genesOut}
\alias{adjust.zstat.in.genesOut}
\title{Adjust MAGMA Z-statistic from .genes.out files}
\usage{
adjust.zstat.in.genesOut(
  ctd,
 
    magma_file = "/Users/natske/GWAS_Summary_Statistics/MAGMA_Files/20016.assoc.tsv.10UP.1.5DOWN/20016.assoc.tsv.10UP.1.5DOWN.genes.out",
  sctSpecies = "mouse"
)
}
\arguments{
\item{ctd}{Cell type data structure}

\item{magma_file}{A MAGMA .genes.out file}

\item{sctSpecies}{Either 'human' or 'mouse'}
}
\description{
Used when you want to directly analyse the gene level z-scores corrected for gene length etc
}
\examples{
magmaGenesOut = adjust.zstat.in.genesOut(ctd,magma_file="/Users/natske/GWAS_Summary_Statistics/MAGMA_Files/20016.assoc.tsv.10UP.1.5DOWN/20016.assoc.tsv.10UP.1.5DOWN.genes.out",sctSpecies="mouse")

}
