% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_zstat_in_genesOut.r
\name{adjust_zstat_in_genesOut}
\alias{adjust_zstat_in_genesOut}
\title{Adjust MAGMA Z-statistic from \emph{.genes.out} files}
\usage{
adjust_zstat_in_genesOut(
  magma_GenesOut_file,
  ctd = NULL,
  ctd_species = infer_ctd_species(ctd),
  prepare_ctd = TRUE,
  method = "bonferroni",
  verbose = TRUE,
  ...
)
}
\arguments{
\item{magma_GenesOut_file}{A MAGMA \emph{.genes.out} file 
generated by \link[MAGMA.Celltyping]{map_snps_to_genes}.}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}

\item{prepare_ctd}{Whether to run
\link[MAGMA.Celltyping]{prepare_quantile_groups} on the \code{ctd} first.}

\item{method}{R package to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives consensus scores for each gene mapping based on a
 several different data sources.}
}}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[EWCE:standardise_ctd]{EWCE::standardise_ctd}}
  \describe{
    \item{\code{dataset}}{CellTypeData. name.}
    \item{\code{input_species}}{Which species the gene names in \code{exp} come from.
See \link[EWCE]{list_species} for all available species.}
    \item{\code{output_species}}{Which species' genes names to convert \code{exp} to.
See \link[EWCE]{list_species} for all available species.}
    \item{\code{non121_strategy}}{How to handle genes that don't have
1:1 mappings between \code{input_species}:\code{output_species}.
Options include:\cr
\itemize{
\item{\code{"drop_both_species" or "dbs" or 1} : \cr}{
Drop genes that have duplicate
mappings in either the \code{input_species} or \code{output_species} \cr
(\emph{DEFAULT}).}
\item{\code{"drop_input_species" or "dis" or 2} : \cr}{
Only drop genes that have duplicate
mappings in the \code{input_species}.}
\item{\code{"drop_output_species" or "dos" or 3} : \cr}{
Only drop genes that have duplicate
mappings in the \code{output_species}.}
\item{\code{"keep_both_species" or "kbs" or 4} : \cr}{
Keep all genes regardless of whether
they have duplicate mappings in either species.}
\item{\code{"keep_popular" or "kp" or 5} : \cr}{
Return only the most "popular" interspecies ortholog mappings.
 This procedure tends to yield a greater number of returned genes
 but at the cost of many of them not being true biological 1:1 orthologs.}
 \item{\code{"sum","mean","median","min" or "max"} : \cr}{
 When \code{gene_df} is a matrix and \code{gene_output="rownames"},
  these options will aggregate many-to-one gene mappings
  (\code{input_species}-to-\code{output_species})
  after dropping any duplicate genes in the \code{output_species}.
 }
}}
    \item{\code{force_new_quantiles}}{By default, quantile computation is
skipped if they have already been computed.
Set \code{=TRUE} to override this and generate new quantiles.}
    \item{\code{force_standardise}}{If \code{ctd} has already been standardised, whether
to rerun standardisation anyway (Default: \code{FALSE}).}
    \item{\code{remove_unlabeled_clusters}}{Remove any samples that have
numeric column names.}
    \item{\code{numberOfBins}}{Number of non-zero quantile bins.}
    \item{\code{keep_annot}}{Keep the column annotation data if provided.}
    \item{\code{keep_plots}}{Keep the dendrograms if provided.}
    \item{\code{as_sparse}}{Convert to sparse matrix.}
    \item{\code{as_DelayedArray}}{Convert to \code{DelayedArray}.}
    \item{\code{rename_columns}}{Remove \code{replace_chars} from column names.}
    \item{\code{make_columns_unique}}{Rename each columns with the prefix
\code{dataset.species.celltype}.}
  }}
}
\description{
Used when you want to directly analyse the gene-level
 Z-scores for a given GWAS while correcting 
 for known confounding variables such as: 
\itemize{
\item{\code{NSNPS} : }{Number of SNPs}
\item{\code{NPARAM} : }{Number of parameters?}
\item{\code{GENELEN} : }{Gene length}
\item{\code{log***} : }{The logged version of each of the above variables, 
using the default \link[base]{log} function.}
}
}
\examples{
myGenesOut <- MAGMA.Celltyping::import_magma_files(
    ids = c("ieu-a-298"),
    file_types = ".genes.out",
    return_dir = FALSE)
ctd <- ewceData::ctd()

magmaGenesOut <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd,
    magma_GenesOut_file = myGenesOut,
    ctd_species = "mouse"
)
}
