% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_hgnc_wtEntrez}
\alias{all_hgnc_wtEntrez}
\title{All HGNC gene symbols with ENTREZ gene IDs}
\format{An object of class \code{data.frame} with 38923 rows and 2 columns.}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
library("biomaRt")
listMarts(host="www.ensembl.org")
human <- useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl")
all_hgnc_wtEntrez = getBM(attributes=c("hgnc_symbol","entrezgene"), mart=human)
devtools::use_data(all_hgnc_wtEntrez)
}
}
\usage{
all_hgnc_wtEntrez
}
\description{
A dataset containing all HGNC symbols in first column, then entrez in second column
}
\keyword{datasets}
