% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_snp_location_tables.r
\name{build_snp_location_tables}
\alias{build_snp_location_tables}
\title{Build data table with RSID / CHR / BP / GenomeBuild for all Hapmap3 SNPs in GRCh37 and GRCh38}
\usage{
build_snp_location_tables()
}
\value{
Save SNP_LOC_DATA to package data
}
\description{
Build data table with RSID / CHR / BP / GenomeBuild for all Hapmap3 SNPs in GRCh37 and GRCh38
}
\examples{
# build_snp_location_tables()

# @import R.utils
}
