% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate.celltype.enrichment.probabilities.wtLimma.r
\name{calculate.celltype.enrichment.probabilities.wtLimma}
\alias{calculate.celltype.enrichment.probabilities.wtLimma}
\title{Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files}
\usage{
calculate.celltype.enrichment.probabilities.wtLimma(
  magmaAdjZ,
  ctd,
  thresh = 1e-04,
  sctSpecies = "mouse"
)
}
\arguments{
\item{magmaAdjZ}{Output from adjust.zstat.in.genesOut()}

\item{ctd}{Cell type data structure}

\item{thresh}{A threshold on low specificity values (used to drop genes)}

\item{sctSpecies}{Either 'human' or 'mouse'}
}
\description{
Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files
}
\examples{
# The package stores an example genesOut file, so save this to a tempfile
myGenesOut = tempfile()
data.table::fwrite(MAGMA.Celltyping::genesOut,sep="\t",file=myGenesOut)
magmaGenesOut = adjust.zstat.in.genesOut(EWCE::ctd,magma_GenesOut_file=myGenesOut,sctSpecies="mouse")

}
