% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate.conditional.celltype.enrichment.probabilities.wtLimma.r
\name{calculate.conditional.celltype.enrichment.probabilities.wtLimma}
\alias{calculate.conditional.celltype.enrichment.probabilities.wtLimma}
\title{Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files}
\usage{
calculate.conditional.celltype.enrichment.probabilities.wtLimma(
  magma1,
  magma2,
  ctd,
  thresh = 1e-04,
  sctSpecies = "mouse",
  annotLevel = 1
)
}
\arguments{
\item{magma1}{Output from adjust.zstat.in.genesOut() (Alzheimers for instance)}

\item{magma2}{Output from adjust.zstat.in.genesOut() (EduYears for instance)}

\item{ctd}{Cell type data structure}

\item{thresh}{A threshold on low specificity values (used to drop genes)}

\item{sctSpecies}{Either 'human' or 'mouse'}

\item{annotLevel}{Annotation level of CellTypeData to use (integer)}
}
\description{
Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files
}
\examples{
magmaGenesOut = adjust.zstat.in.genesOut(ctd,magma_file="/Users/natske/GWAS_Summary_Statistics/MAGMA_Files/20016.assoc.tsv.10UP.1.5DOWN/20016.assoc.tsv.10UP.1.5DOWN.genes.out",sctSpecies="mouse")

}
