% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_celltype_enrichment_limma.r
\name{calculate_celltype_enrichment_limma}
\alias{calculate_celltype_enrichment_limma}
\title{Calculate celltype enrichment}
\usage{
calculate_celltype_enrichment_limma(
  magmaAdjZ,
  ctd,
  ctd_species = "mouse",
  annotLevel = 1,
  prepare_ctd = TRUE,
  thresh = 1e-04,
  celltypes = NULL,
  return_all = FALSE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{magmaAdjZ}{Output from
\link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}.}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}

\item{annotLevel}{Level within the \code{ctd} to use.}

\item{prepare_ctd}{Whether to run
\link[MAGMA.Celltyping]{prepare_quantile_groups} on the \code{ctd} first.}

\item{thresh}{A threshold on low specificity values (used to drop genes).}

\item{celltypes}{Cell types to conduct tests for. Defaults to all cell types
available in selected \code{ctd} level.}

\item{return_all}{Return a list of both the processed 
data input and the enrichment results (Default).
 If \code{FALSE}, will only return enrichment results.}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[EWCE]{standardise_ctd}.}
}
\description{
Calculate celltype enrichments without MAGMA
using adjusted MAGMA Z-statistic from .genes.out files.
}
\examples{
ctd <- ewceData::ctd()

# The package stores an example genesOut file, so save this to a tempfile
myGenesOut <- MAGMA.Celltyping::import_magma_files(
    ids = c("ieu-a-298"),
    file_types = ".genes.out",
    return_dir = FALSE)
    
#### Get adjust gene z-scores ####    
magmaAdjZ <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd,
    magma_GenesOut_file = myGenesOut)
    
#### Get cell type enrichment ####
ps <- MAGMA.Celltyping::calculate_celltype_enrichment_limma(
    magmaAdjZ = magmaAdjZ,
    ctd = ctd)
}
