% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_conditional_celltype_enrichment_limma.R
\name{calculate_conditional_celltype_enrichment_limma}
\alias{calculate_conditional_celltype_enrichment_limma}
\title{Calculate celltype enrichments without MAGMA}
\usage{
calculate_conditional_celltype_enrichment_limma(
  magma1,
  magma2,
  ctd,
  ctd_species = infer_ctd_species(ctd),
  annotLevel = 1,
  prepare_ctd = TRUE,
  thresh = 1e-04,
  show_plot = TRUE,
  verbose = TRUE,
  ...
)
}
\arguments{
\item{magma1}{Output from \link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}
(e.g. trait 1).}

\item{magma2}{Output from \link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}
(e.g. trait 2).}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}

\item{annotLevel}{Level within the \code{ctd} to use.}

\item{prepare_ctd}{Whether to run
\link[MAGMA.Celltyping]{prepare_quantile_groups} on the \code{ctd} first.}

\item{thresh}{A threshold on low specificity values (used to drop genes).}

\item{show_plot}{Show plot visualizations of results (Default: \code{TRUE}).}

\item{verbose}{Print messages.}

\item{...}{Additional arguments passed to \link[EWCE]{standardise_ctd}.}
}
\value{
Data.frame with conditional enrichment results.
}
\description{
Calculate celltype enrichments without MAGMA
using adjusted MAGMA Z-statistic from .genes.out files.
}
\examples{
ctd <- ewceData::ctd()

#### MAGMA 1 #####
genesOut1 <- MAGMA.Celltyping::import_magma_files(
    ids = c("ieu-a-298"),
    file_types = ".genes.out",
    return_dir = FALSE)
magma1 <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd, 
    magma_GenesOut_file = genesOut1)

#### MAGMA 2 #####
genesOut2 <- MAGMA.Celltyping::import_magma_files(
    ids = c("ukb-b-6548"),
    file_types = ".genes.out",
    return_dir = FALSE)
magma2 <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd, 
    magma_GenesOut_file = genesOut2)

#### Run conditional enrichment ####
cond_res <- 
MAGMA.Celltyping::calculate_conditional_celltype_enrichment_limma(
    magma1 = magma1,
    magma2 = magma2,
    ctd = ctd)
}
