% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_conditional_celltype_enrichment_probabilities_wtLimma.R
\name{calculate_conditional_celltype_enrichment_probabilities_wtLimma}
\alias{calculate_conditional_celltype_enrichment_probabilities_wtLimma}
\title{Calculate celltype enrichments without MAGMA}
\usage{
calculate_conditional_celltype_enrichment_probabilities_wtLimma(
  magma1,
  magma2,
  ctd,
  thresh = 1e-04,
  sctSpecies = "mouse",
  annotLevel = 1
)
}
\arguments{
\item{magma1}{Output from \link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}
(e.g. trait 1).}

\item{magma2}{Output from \link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}
(e.g. trait 2).}

\item{ctd}{Cell type data structure containing $specificity_quantiles.}

\item{thresh}{A threshold on low specificity values (used to drop genes)}

\item{sctSpecies}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}
}
\description{
Calculate celltype enrichments without MAGMA
using adjusted MAGMA Z-statistic from .genes.out files.
}
