% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conditional_geneset_enrichment.R
\name{calculate_conditional_geneset_enrichment}
\alias{calculate_conditional_geneset_enrichment}
\title{Use MAGMA to test enrichment in a geneset}
\usage{
calculate_conditional_geneset_enrichment(
  geneset,
  ctd,
  ctd_species = "mouse",
  prepare_ctd = TRUE,
  controlledAnnotLevel = 1,
  controlledCTs,
  gwas_sumstats_path = NULL,
  magma_dir = NULL,
  analysis_name = "MainRun",
  upstream_kb = 35,
  downstream_kb = 10,
  genome_ref_path = NULL,
  geneset_species = "mouse",
  verbose = TRUE
)
}
\arguments{
\item{geneset}{Genes which are to be tested (as HGNC / MGI symbols).}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}

\item{prepare_ctd}{Whether to run
\link[MAGMA.Celltyping]{prepare_quantile_groups} on the \code{ctd} first.}

\item{controlledAnnotLevel}{Annotation level of the celltypes
being controlled for.}

\item{controlledCTs}{Array of the celltype to be controlled for,
e.g. c("Interneuron type 16","Medium Spiny Neuron).}

\item{gwas_sumstats_path}{File path of the summary statistics file.}

\item{analysis_name}{Used in file names which area created.}

\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}

\item{genome_ref_path}{Path to the folder containing the 1000
genomes reference (downloaded with \link[MAGMA.Celltyping]{get_genome_ref}).}

\item{geneset_species}{Species name relevant to the genes in the geneset.}

\item{verbose}{Print messages.}
}
\value{
File path for the genes.out file.
}
\description{
Assumes that you have already run \link[MAGMA.Celltyping]{map_snps_to_genes}.
}
\examples{
#### Import data ####
ctd <- MAGMA.Celltyping::get_ctd("ctd_allKI")
magma_dir <- MAGMA.Celltyping::import_magma_files(ids = "ieu-a-298")
geneset <- MAGMA.Celltyping::rbfox_binding

res <- MAGMA.Celltyping::calculate_conditional_geneset_enrichment(
    geneset = geneset,
    ctd = ctd,
    controlledAnnotLevel = 1,
    controlledCTs = "pyramidal SS",
    magma_dir = magma_dir,
    analysis_name = "Rbfox_16_pyrSS",  
    geneset_species = "mouse",
    ctd_species = "mouse")

}
