% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_geneset_enrichment.r
\name{calculate_geneset_enrichment}
\alias{calculate_geneset_enrichment}
\title{Use MAGMA to test enrichment in a geneset}
\usage{
calculate_geneset_enrichment(
  geneset,
  gwas_sumstats_path = NULL,
  magma_dir = NULL,
  analysis_name,
  upstream_kb = 35,
  downstream_kb = 10,
  geneset_species = "mouse",
  version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{geneset}{Genes which are to be tested. 
Can be gene symbols (or other gene IDs) from any species listed in 
\link[EWCE]{list_species}.}

\item{gwas_sumstats_path}{File path of the summary statistics file.}

\item{magma_dir}{Path to folder containing the
pre-computed MAGMA GWAS files (\emph{.gsa.raw}and \emph{.gsa.out}).}

\item{analysis_name}{Used in file names which area created.}

\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}

\item{geneset_species}{Species that the \code{geneset} came from.
Can be any species listed in \link[EWCE]{list_species}.
If \code{geneset_species!="human"}, the \code{geneset} will be converted
 to 1:1 human orthologs using \link[orthogene]{convert_orthologs}.}

\item{version}{MAGMA version to use.}

\item{verbose}{Print messages.}
}
\value{
Filepath for the genes.out file.
}
\description{
Assumes that you have already run
\link[MAGMA.Celltyping]{map_snps_to_genes}.
}
