% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/capture.R
\name{capture}
\alias{capture}
\title{Capture output of function}
\usage{
capture(func_output)
}
\arguments{
\item{func_output}{Function output.}
}
\value{
Null output.
}
\description{
Capture output of function and print to screen as a message.
This allows tables, summaries, etc to be printed to the console without 
messing up alignment of columns.
}
\keyword{internal}
