% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_associations_pipeline.r
\name{celltype_associations_pipeline}
\alias{celltype_associations_pipeline}
\title{Calculate cell type associations using MAGMA}
\usage{
celltype_associations_pipeline(
  ctd,
  ctd_levels = seq_len(length(ctd)),
  ctd_name,
  ctd_species = infer_ctd_species(ctd),
  standardise = TRUE,
  magma_dirs,
  run_linear = TRUE,
  run_top10 = TRUE,
  run_conditional = FALSE,
  upstream_kb = 35,
  downstream_kb = 10,
  suffix_linear = "linear",
  suffix_top10 = "top10",
  suffix_condition = "condition",
  controlledAnnotLevel = 1,
  controlTopNcells = 1,
  force_new = FALSE,
  save_dir = tempdir(),
  nThread = 1,
  version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_levels}{Which levels of \code{ctd} to
iterate the enrichment analysis over.}

\item{ctd_name}{Used in file names}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}

\item{standardise}{Whether to run \link[EWCE]{standardise_ctd} first.
Provides gene ortholog conversion.}

\item{magma_dirs}{Path to folders containing the
pre-computed MAGMA GWAS files (\emph{.gsa.raw}and \emph{.gsa.out}).
\emph{NOTE}: Files within these folders must have the same naming scheme
as the folders themselves.}

\item{run_linear}{Run in linear mode.}

\item{run_top10}{Run  in top 10\% mode.}

\item{run_conditional}{Run in conditional mode.}

\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}

\item{suffix_linear}{This will be added to the linear results file name.}

\item{suffix_top10}{This will be added to the top 10\% results file name.}

\item{suffix_condition}{This will be added to the
conditional results file name.}

\item{controlledAnnotLevel}{Which annotation level should be controlled for.}

\item{controlTopNcells}{How many of the most significant cell types at
that annotation level should be controlled for?}

\item{force_new}{[Optional] Force new MAGMA analyses even if the
pre-existing results files are detected.}

\item{save_dir}{Folder to save results in (\code{save_dir=NULL}
to not save any results).}

\item{nThread}{Number of threads to parallelise analyses across.}

\item{version}{MAGMA version to use.}

\item{verbose}{Print messages.}
}
\value{
A list containing the results of each selected
celltype associations analysis.
}
\description{
Has the option of running multiple analyses with a single function.
Assumes that you already have all MAGMA GWAS files precomputed.
Precomputed MAGMA GWAS files can be downloaded via the
\link[MAGMA.Celltyping]{import_magma_files} function.
}
\examples{
magma_dirs <- MAGMA.Celltyping::import_magma_files(ids = c("ieu-a-298"))
ctd <- ewceData::ctd()

res <- MAGMA.Celltyping::celltype_associations_pipeline(
    ctd = ctd,
    ctd_levels = 1,
    ctd_name = "Zeisel2015",
    ctd_species = "mouse",
    magma_dirs = magma_dirs)
}
\keyword{main_function}
