% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_access.R
\name{check_access}
\alias{check_access}
\title{Check access}
\usage{
check_access(dest_dir)
}
\arguments{
\item{dest_dir}{Directory to check.}
}
\description{
Check if one has access to directory
}
\keyword{internal}
