% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_enrichment_mode.R
\name{check_enrichment_mode}
\alias{check_enrichment_mode}
\title{Check enrichment mode}
\usage{
check_enrichment_mode(EnrichmentMode)
}
\description{
For functions where \code{EnrichmentMode} is an argument, 
check whether a valid option has been selected.
}
\keyword{internal}
