% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_entrez_genes.R
\name{check_entrez_genes}
\alias{check_entrez_genes}
\title{Check ENTREZ genes}
\usage{
check_entrez_genes(geneset)
}
\description{
Check percentage of gene genes in \code{geneset} that have 
corresponding ENTREZ IDs.
}
\keyword{internal}
