% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_inputs_to_magma_celltype_analysis.r
\name{check_inputs_to_magma_celltype_analysis}
\alias{check_inputs_to_magma_celltype_analysis}
\title{Check input to MAGMA cell type analyses}
\usage{
check_inputs_to_magma_celltype_analysis(
  ctd,
  gwas_sumstats_path,
  upstream_kb,
  downstream_kb
)
}
\arguments{
\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{gwas_sumstats_path}{File path of the summary statistics file.}

\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}
}
\value{
Nothing returned but sends error if there is a fault
with the arguments.
}
\description{
This code was functonalised because it is called by both baseline
and conditional analysis.
}
\keyword{internal}
