% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_quantiles.R
\name{check_quantiles}
\alias{check_quantiles}
\title{Check quantiles}
\usage{
check_quantiles(ctd, matrix_name, numberOfBins, metric = c("n", "max"))
}
\description{
Check whether every column in every matrix in every level of the 
CellTypeDataset (\code{ctd}) has the expected number of quantile bins.
}
\keyword{internal}
