% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/choose_ewce_bin_function.R
\name{choose_ewce_bin_function}
\alias{choose_ewce_bin_function}
\title{Chooses EWCE function}
\usage{
choose_ewce_bin_function(func = "bin_columns_into_quantiles")
}
\value{
Function.
}
\description{
Identifies the name of the EWCE binning 
function to use depending on which version 
of \pkg{EWCE} is currently installed.
}
\keyword{internal}
