% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.trait.enrichment.r
\name{compare.trait.enrichments}
\alias{compare.trait.enrichments}
\title{Test to see which cell types are more enriched in magmaPath2 than magmaPath1}
\usage{
compare.trait.enrichments(
  magmaPath1 = NA,
  magmaPath2 = NA,
  magma1 = NA,
  magma2 = NA,
  annotLevel,
  ctd
)
}
\arguments{
\item{magmaPath1}{Filepath of magma file}

\item{magmaPath2}{Filepath of magma file}

\item{magma1}{Results of MAGMA analysis}

\item{magma2}{Results of MAGMA analysis}

\item{annotLevel}{At which annotation level of the CTD should the comparison be done?}

\item{ctd}{Cell type data strucutre containing $specificity_quantiles}
}
\value{
Filepath for the genes.out file
}
\description{
Test to see which cell types are more enriched in magmaPath2 than magmaPath1
}
