% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_trait_enrichment.r
\name{compare_trait_enrichments}
\alias{compare_trait_enrichments}
\title{Test to see which cell types are more enriched in magmaPath2 than magmaPath1}
\usage{
compare_trait_enrichments(
  magmaPath1 = NA,
  magmaPath2 = NA,
  magma1 = NA,
  magma2 = NA,
  annotLevel,
  ctd
)
}
\arguments{
\item{magmaPath1}{A magma results file}

\item{magmaPath2}{A magma results file}

\item{magma1}{A magma results data structure}

\item{magma2}{A magma results data structure}

\item{annotLevel}{Annotation level of the CTD / results file to use}

\item{ctd}{Cell type data structure containing $specificity_quantiles}
}
\value{
File path for the genes.out file
}
\description{
Test to see which cell types are more enriched in magmaPath2 than magmaPath1
}
