% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copy_snps_to_genes_files.R
\name{copy_snps_to_genes_files}
\alias{copy_snps_to_genes_files}
\title{Copy MAGMA gene mapping files to new folder}
\usage{
copy_snps_to_genes_files(
  gene_files,
  save_dir = "MAGMA_Files",
  overwrite = FALSE
)
}
\arguments{
\item{gene_files}{Full paths to gene mapping mapping files.}

\item{save_dir}{Where to save the gene mapping files to.}

\item{overwrite}{Whether to overwrite gene mapping files that already
exist in \code{save_dir} (Default: \code{FALSE}).}
}
\value{
Named character list
}
\description{
Copy SNPs-to-genes mapping files (\emph{.genes.out}) produced by 
\link[MAGMA.Celltyping]{map_snps_to_genes} to a new folder. 
Importantly, this function maintains the file naming and folder structure 
conventions necessary to run other downstream 
\pkg{MAGMA.Celltyping} functions.
}
\keyword{internal}
