% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_fake_gwas_path.R
\name{create_fake_gwas_path}
\alias{create_fake_gwas_path}
\title{Create a fake GWAS path when all you have are the MAGMA files}
\usage{
create_fake_gwas_path(
  magma_dir,
  upstream_kb = 35,
  downstream_kb = 10,
  remove_pattern = "/MAGMA_Files/"
)
}
\arguments{
\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}
}
\description{
Necessary to trick 
\link[MAGMA.Celltyping]{calculate_celltype_associations} and 
\link[MAGMA.Celltyping]{calculate_conditional_celltype_associations} to 
take MAGMA files (".genes.raw" and ".genes.out") alone.
}
\keyword{internal}
