% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_covar_file.r
\name{create_gene_covar_file}
\alias{create_gene_covar_file}
\title{Create gene covariance file}
\source{
\code{
#### Example usage ####
ctd <- ewceData::ctd()
genesOutFile <- MAGMA.Celltyping::import_magma_files(
    ids = "ieu-a-298",
    file_types = ".genes.out$",
    return_dir = FALSE)

genesCovarFilePath <- create_gene_covar_file(
    genesOutFile = genesOutFile,
    ctd = ctd,
    annotLevel = 1,
    ctd_species = "mouse")
}
}
\usage{
create_gene_covar_file(
  genesOutFile,
  ctd,
  annotLevel,
  ctd_species,
  genesOutCOND = NA
)
}
\arguments{
\item{genesOutFile}{The output of the second call to MAGMA
(performed in the \link[MAGMA.Celltyping]{map_snps_to_genes} function).}

\item{ctd}{Cell type data structure. Must contain quantiles.}

\item{annotLevel}{Annot level for which the gene covar file
should be constructed}

\item{ctd_species}{Species name relevant to the cell type data,
i.e. "mouse" or "human"}

\item{genesOutCOND}{[Optional] Path to a \emph{genes.out}
file to condition on. Used if you want to condition on a different GWAS.}
}
\value{
File path for the gene covar file.
}
\description{
The gene covariannce (covar) file is the input to MAGMA for the
celltype association analysis.
This code was functonalised because it is called by both
baseline and conditional analysis.
}
\keyword{internal}
