% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_gene_covar_file_mode.R
\name{create_gene_covar_file_mode}
\alias{create_gene_covar_file_mode}
\title{Create gene covariance file: by mode}
\usage{
create_gene_covar_file_mode(
  EnrichmentMode,
  magmaPaths,
  ctd,
  annotLevel,
  ctd_species
)
}
\arguments{
\item{EnrichmentMode}{[Optional] Should either 'Linear' or 'Top 10\%' mode
be used for testing enrichment?}

\item{magmaPaths}{Output of \code{get_magma_paths}.}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}
}
\description{
Create gene covariance file for a given \code{EnrichmentMode}.
}
\keyword{internal}
