% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_top10percent_genesets_file.r
\name{create_top10percent_genesets_file}
\alias{create_top10percent_genesets_file}
\title{Create gene covar file}
\usage{
create_top10percent_genesets_file(
  genesOutFile,
  ctd,
  annotLevel,
  sctSpecies,
  verbose = TRUE
)
}
\arguments{
\item{genesOutFile}{The output of the second call to MAGMA
(performed in the map_snps_to_genes function).}

\item{ctd}{Cell type data structure containing $specificity_quantiles.}

\item{sctSpecies}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}

\item{verbose}{Print messages.}
}
\value{
Filepath for the gene covar file
}
\description{
The gene covar file is the input to MAGMA for the celltype
association analysis. This code was functonalised
because it is called by both baseline and conditional analysis.
}
\examples{
myGenesOut <- tempfile()
ctd <- ewceData::ctd()
data.table::fwrite(
    x = MAGMA.Celltyping::genesOut,
    file = myGenesOut,
    sep = "\t"
)
ctd <- MAGMA.Celltyping::prepare_quantile_groups(ctd = ctd)
geneSetsFilePath <- MAGMA.Celltyping::create_top10percent_genesets_file(
    genesOutFile = myGenesOut,
    ctd = ctd,
    annotLevel = 1,
    sctSpecies = "mouse"
)
}
\keyword{internal}
