% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_AIBS}
\alias{ctd_AIBS}
\title{Celltype data (AIBS)}
\format{An object of class \code{list} of length 1.}
\source{
# http://www.hjerling-leffler-lab.org/data/scz_singlecell/
usethis::use_data(ctd_AIBS,overwrite=TRUE)
}
\usage{
ctd_AIBS
}
\description{
CTD file with AIBS human cortex data
}
\keyword{datasets}
