% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_allKI}
\alias{ctd_allKI}
\title{RSD / CHR / BP for GRCh37 and GRCh38}
\format{An object of class \code{list} of length 2.}
\source{
build_snp_location_tables()

Celltype data (allKI)

CTD file with cortex, hippocampus, hypothalamus and midbrain

# http://www.hjerling-leffler-lab.org/data/scz_singlecell/
usethis::use_data(ctd_allKI,overwrite=TRUE)
}
\usage{
ctd_allKI
}
\description{
Data table with all Hapmap3 SNPs
}
\keyword{datasets}
