% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_BlueLake2018_VisualCortexOnly}
\alias{ctd_BlueLake2018_VisualCortexOnly}
\title{Celltype data (Blue Lake 2018 Visual Cortex)}
\format{
An object of class \code{list} of length 2.
}
\source{
# Generated on Nathan's Mac: ~/Single Cell Datasets/BlueLake2018
usethis::use_data(ctd_BlueLake2018_FrontalCortexOnly,overwrite=TRUE)
}
\usage{
ctd_BlueLake2018_VisualCortexOnly
}
\description{
CTD file with data from humans
}
\keyword{datasets}
