% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_DivSeq}
\alias{ctd_DivSeq}
\title{Celltype data (DivSeq)}
\format{
An object of class \code{list} of length 2.
}
\source{
usethis::use_data(ctd_DivSeq,overwrite = TRUE)
}
\usage{
ctd_DivSeq
}
\description{
CTD file with DivSeq data
}
\keyword{datasets}
