% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_DRONC_human}
\alias{ctd_DRONC_human}
\title{Celltype data (DRONC_human)}
\format{An object of class \code{list} of length 2.}
\source{
# http://www.hjerling-leffler-lab.org/data/scz_singlecell/
usethis::use_data(ctd_DRONC_human,overwrite=TRUE)
}
\usage{
ctd_DRONC_human
}
\description{
CTD file with DRONC-seq data from human
}
\keyword{datasets}
