% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_Saunders}
\alias{ctd_Saunders}
\title{Celltype data (Saunders)}
\format{An object of class \code{list} of length 2.}
\source{
# Generated on Nathan's Mac: ctd_SaundersWithLvl1.Rda
usethis::use_data(ctd_Saunders,overwrite=TRUE)
}
\usage{
ctd_Saunders
}
\description{
CTD file with data from humans
}
\keyword{datasets}
