% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_Tasic}
\alias{ctd_Tasic}
\title{Celltype data (Tasic)}
\format{
An object of class \code{list} of length 2.
}
\source{
usethis::use_data(ctd_Tasic,overwrite = TRUE)
}
\usage{
ctd_Tasic
}
\description{
CTD file with Tasic data
}
\keyword{datasets}
