% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_Zeisel2018}
\alias{ctd_Zeisel2018}
\title{Celltype data (Zeisel2018)}
\format{
An object of class \code{list} of length 5.
}
\source{
# Generated on Nathan's Mac: ctd_ZieselAllLevels.Rda
usethis::use_data(ctd_Zeisel2018,overwrite=TRUE)
}
\usage{
ctd_Zeisel2018
}
\description{
CTD file with data from mice
}
\keyword{datasets}
