% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop.genes.within.mhc.r
\name{drop.genes.within.mhc}
\alias{drop.genes.within.mhc}
\title{Drop genes within MHC}
\usage{
drop.genes.within.mhc(geneLocFilePath, geneLocFilePathOut)
}
\arguments{
\item{geneLocFilePath}{The gene loc file as downloaded from
MAGMA website (e.g. NCBI37.3.gene.loc).}

\item{geneLocFilePathOut}{The file path to write the gene
loc file with MHC removed.}
}
\value{
NULL (the output is written to the filepath provided as an argument)
}
\description{
Drop genes within MHC
}
\examples{
drop.genes.within.mhc("data/NCBI37.3.gene.loc", "data/NCBI37.3.gene.loc")
drop.genes.within.mhc("data/NCBI38.gene.loc", "data/NCBI38.gene.loc")
drop.genes.within.mhc("data/NCBI36.3.gene.loc", "data/NCBI36.3.gene.loc")
}
