% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drop_genes_within_mhc.r
\name{drop_genes_within_mhc}
\alias{drop_genes_within_mhc}
\title{Drop genes within MHC}
\usage{
drop_genes_within_mhc(geneLocFilePath, geneLocFilePathOut)
}
\arguments{
\item{geneLocFilePath}{The gene loc file as downloaded from
MAGMA website (e.g. NCBI37.3.gene.loc).}

\item{geneLocFilePathOut}{The file path to write the gene
loc file with MHC removed.}
}
\value{
NULL (the output is written to the filepath provided as an argument)
}
\description{
Drop genes within MHC
}
\examples{
drop_genes_within_mhc("data/NCBI37.3.gene.loc", "data/NCBI37.3.gene.loc")
drop_genes_within_mhc("data/NCBI38.gene.loc", "data/NCBI38.gene.loc")
drop_genes_within_mhc("data/NCBI36.3.gene.loc", "data/NCBI36.3.gene.loc")
}
