% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{enrichment_results}
\alias{enrichment_results}
\title{Example celltype enrichment results}
\format{
An object of class \code{list} of length 2.
}
\source{
\code{ 
#### Import data ####
magma_dirs <- MAGMA.Celltyping::import_magma_files(ids = "ieu-a-298")
ctd <- ewceData::ctd()
enrichment_results <- MAGMA.Celltyping::celltype_associations_pipeline(
    magma_dirs = magma_dirs,
    ctd = ctd, 
    ctd_species = "mouse", 
    ctd_name = "Zeisel2015", 
    run_linear = TRUE, 
    run_top10 = TRUE)
usethis::use_data(enrichment_results, overwrite = TRUE)
}
}
\usage{
data("enrichment_results")
}
\description{
Example results from \link[MAGMA.Celltyping]{celltype_associations_pipeline}.
}
\keyword{datasets}
