% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.by.orthologs.r
\name{filter.by.orthologs}
\alias{filter.by.orthologs}
\title{Filter specificity and mean_exp from CTD list to only keep a subset of genes}
\usage{
\method{filter}{by.orthologs}(ctd_oneLevel, one2one_ortholog_symbols)
}
\arguments{
\item{ctd_oneLevel}{method to order colors (\code{"hsv"} or \code{"cluster"})}

\item{one2one_ortholog_symbols}{character expansion for the text}
}
\value{
The ctd with both specificity and mean expression matrices filtered to only contain 1:1 ortholog genes
}
\description{
Created so that lapply can be used to do this
}
\examples{
ctd2 = lapply(ctd,filter.by.orthologs,one2one_ortholog_symbols = ortholog_data[,2])

}
