% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_by_orthologs.r
\name{filter_by_orthologs}
\alias{filter_by_orthologs}
\title{Filter CellTypeDataset}
\source{
\code{
ctd <- ewceData::ctd()
orths <- ewceData::mouse_to_human_homologs()
ctd2 <- lapply(ctd, filter_by_orthologs, orths$MGI.symbol)
}
}
\usage{
filter_by_orthologs(ctd_oneLevel, one2one_ortholog_symbols)
}
\arguments{
\item{ctd_oneLevel}{A single level of a CTD}

\item{one2one_ortholog_symbols}{character expansion for the text.}
}
\value{
The CTD with both specificity and mean
 expression matrices filtered to only contain 1:1 ortholog genes.
}
\description{
Filter CellTypeDataset (CTD) 
specificity and mean_exp from CTD list to only keep a subset of genes.
}
\details{
Created so that lapply can be used to do this.
}
\keyword{internal}
