% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_GenesOut_files.R
\name{find_GenesOut_files}
\alias{find_GenesOut_files}
\title{Find genes.out files}
\usage{
find_GenesOut_files(GenesOut_dir, verbose = TRUE)
}
\arguments{
\item{GenesOut_dir}{Directory to search in.}

\item{verbose}{Print messages.}
}
\description{
Search for MAGMA output files (genes.out) in a given directory.
Support function for \link[MAGMA.Celltyping]{get_driver_genes}.
}
\keyword{internal}
