% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_install_dir.R
\name{find_install_dir}
\alias{find_install_dir}
\title{Find a directory to install software in}
\usage{
find_install_dir(
  dest_dir_opts = c(tools::R_user_dir("MAGMA.Celltyping", which = "cache"),
    system.file("tools", package = "MAGMA.Celltyping"), "/usr/local/bin",
    Sys.getenv("HOME"), getwd(), tempdir()),
  verbose = TRUE
)
}
\arguments{
\item{dest_dir_opts}{Potential installation directory options.}

\item{verbose}{Print messages.}
}
\value{
Path to viable installation directory.
}
\description{
Find a directory that you have write permissions in to so that you can 
install software there. 
Tests several options and returns the first viable one.
}
\keyword{internal}
