% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fix_path.R
\name{fix_path}
\alias{fix_path}
\title{Fix path}
\usage{
fix_path(x)
}
\arguments{
\item{x}{File/directory path.}
}
\description{
Ensure a file/directory path can be understood by all OS types by expanding 
"~" to an absolute path, and standardized slash directions with 
\link[base]{normalizePath}. 
This is especially important for Windows.
}
\keyword{internal}
