% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.sumstats.for.magma.macOnly.R
\name{format.sumstats.for.magma.macOnly}
\alias{format.sumstats.for.magma.macOnly}
\title{Check that sumstats has correct columns and that they are
 in the correct order for MAGMA and LDSC.}
\usage{
\method{format}{sumstats.for.magma.macOnly}(path)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted.}
}
\value{
col_headers The new column headers for the sumstats file
}
\description{
Check that sumstats has correct columns and that they are
 in the correct order for MAGMA and LDSC.
}
\keyword{internal}
