% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format.sumstats.for.magma.r
\name{format.sumstats.for.magma}
\alias{format.sumstats.for.magma}
\title{Format GWAS summary statistics}
\usage{
\method{format}{sumstats.for.magma}(path)
}
\arguments{
\item{path}{Filepath for the summary statistics file to be formatted.}
}
\value{
col_headers The new column headers for the sumstats file
}
\description{
Check that sumstats has correct columns and that they are
 in the correct order for MAGMA and LDSC.
}
\examples{
path <- MAGMA.Celltyping::get_example_gwas(trait = "educational_attainment")
new_path <- MAGMA.Celltyping::format.sumstats.for.magma(path)

}
