% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{genesOut}
\alias{genesOut}
\title{Example of genesOut file}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 18250 rows and 9 columns.
}
\source{
\code{
genesOut = data.table::fread(
    "/home/nskene/tmp_files/20016.assoc.tsv.10UP.1.5DOWN.genes.out")
usethis::use_data(genesOut,overwrite = TRUE)
}
}
\usage{
data("genesOut")
}
\description{
Obtained from \url{'/Users/natske/OneDrive - Imperial College London/GWAS_Summary_Statistics/MAGMA_Files/20016.assoc.tsv.10UP.1.5DOWN/20016.assoc.tsv.10UP.1.5DOWN.genes.out'}
}
\keyword{datasets}
