% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.magma.paths.r
\name{get.magma.paths}
\alias{get.magma.paths}
\title{Get MAGMA Paths}
\usage{
get.magma.paths(
  gwas_sumstats_path = NA,
  upstream_kb = NA,
  downstream_kb = NA,
  output_path = NA
)
}
\arguments{
\item{gwas_sumstats_path}{Path to the GWAS summary statistics file}

\item{upstream_kb}{How many kb upstream of the gene should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene should SNPs be included?}

\item{output_path}{If being called from magma.tileplot() then gwas_sumstats_path cannot be specified, so instead specificify the base folder where MAGMA_Figures exists}
}
\value{
Filepath for where MAGMA files should be created
}
\description{
Gets paths to folders based on location of the GWAS summary statistics file. This avoids the GWAS sumstats folder becoming a mess.
}
\examples{
magmaPaths = get.magma.paths(gwas_sumstats_path,upstream_kb,downstream_kb)

}
