% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd.R
\name{get_ctd}
\alias{get_ctd}
\title{Import precomputed CellTypeDatasets}
\source{

}
\usage{
get_ctd(
  ctd_name = c("ctd_AIBS", "ctd_allKI", "ctd_BlueLake2018_FrontalCortexOnly",
    "ctd_BlueLake2018_VisualCortexOnly", "ctd_DescartesHuman", "ctd_DivSeq",
    "ctd_DRONC_human", "ctd_DRONC_mouse", "ctd_Saunders", "ctd_Tasic", "ctd_Zeisel2018"),
  storage_dir = tempdir()
)
}
\arguments{
\item{ctd_name}{Name of the CellTypeDatset to import.}

\item{storage_dir}{Folder in which to save the CellTypeDatset.}
}
\value{
CellTypeDataset object
}
\description{
Import CellTypeDatasets (CTD) generated by
\href{https://github.com/NathanSkene/EWCE}{EWCE}.
}
\details{
\itemize{

\item{"ctd_AIBS"}{CTD file with AIBS human cortex data.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.}

\item{"ctd_allKI"}{CTD file with cortex, hippocampus,
hypothalamus and midbrain.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.}

\item{"ctd_BlueLake2018_FrontalCortexOnly"}{
CTD file with data from humans.}

\item{"ctd_BlueLake2018_VisualCortexOnly"}{
CTD file with data from humans.}

\item{"ctd_Saunders"}{
CTD file with data from humans.}

\item{"ctd_DescartesHuman"}{
CTD file with scRNA-seq data from human embryo across multiple organ systems.
\href{https://descartes.brotmanbaty.org/bbi/human-gene-expression-during-development/}{Reference}.
}

\item{"ctd_DRONC_human"}{
CTD file with DRONC-seq data from humans.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.
}

\item{"ctd_DRONC_mouse"}{
CTD file with Tasic data from mouse.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.
}

\item{"ctd_DivSeq"}{
CTD file with DivSeq data.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.
}

\item{"ctd_Tasic"}{
CTD file with Tasic data.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.
}

\item{"ctd_Zeisel2018"}{
CTD file with data from mice.
\href{http://www.hjerling-leffler-lab.org/data/scz_singlecell/}{Reference}.
}

}
}
\examples{
ctd <- MAGMA.Celltyping::get_ctd("ctd_AIBS")
}
