% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deciles_matrix.R
\name{get_deciles_matrix}
\alias{get_deciles_matrix}
\title{Get deciles matrix}
\usage{
get_deciles_matrix(ctd, annotLevel, ctd_species)
}
\arguments{
\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}
}
\description{
Extract the matrix of specificity deciles from a CellTypeDataset object.
Support function for
 \link[MAGMA.Celltyping]{create_top10percent_genesets_file}.
}
\keyword{internal}
