% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_gwas.R
\name{get_example_gwas}
\alias{get_example_gwas}
\title{Import GWAS summary statistics}
\source{
\code{
#### fluid_intelligence ####
gwas_sumstats_path <- MAGMA.Celltyping::get_example_gwas(
    trait = "fluid_intelligence", munged = FALSE)
path_formatted <- MungeSumstats::format_sumstats(
    path=gwas_sumstats_path,
    save_path = tempfile(fileext = ".formatted.tsv.gz"),
    ref_genome ="GRCh37")
ss <- data.table::fread(path_formatted)
ss2 <- ss[MINOR_AF>=.05 & P<.05,]
data.table::fwrite(ss2, path_formatted, sep = "\t")
piggyback::pb_upload(file = path_formatted,
                     name = "fluid_intelligence.ukb.tsv.gz",
                     repo = "neurogenomics/MAGMA_Celltyping",
                     overwrite = TRUE)

#### prospective_memory ####
gwas_sumstats_path <- MAGMA.Celltyping::get_example_gwas(
    trait = "prospective_memory", munged = FALSE)
path_formatted <- MungeSumstats::format_sumstats(
    path=gwas_sumstats_path,
    save_path = tempfile(fileext = ".formatted.tsv.gz"),
    ref_genome ="GRCh37")
ss <- data.table::fread(path_formatted)
ss2 <- ss[MINOR_AF>=.05 & P<.05,]
data.table::fwrite(ss2, path_formatted, sep = "\t")
piggyback::pb_upload(file = path_formatted,
                     name = "prospective_memory.ukb.tsv.gz",
                     repo = "neurogenomics/MAGMA_Celltyping",
                     overwrite = TRUE)
}
}
\usage{
get_example_gwas(
  trait = c("educational_attainment", "fluid_intelligence", "prospective_memory"),
  munged = TRUE,
  storage_dir = tools::R_user_dir(package = "MAGMA.Celltyping", which = "cache"),
  timeout = 60 * 5
)
}
\arguments{
\item{trait}{Which trait to get GWAS summary stats for.}

\item{munged}{Whether to download the raw or pre-munged
version of each GWAS (\emph{Default:} \code{TRUE}).}

\item{storage_dir}{Folder in which to store the GWAS summary stats.}

\item{timeout}{How many seconds to wait before timeout.}
}
\value{
Path to downloaded GWAS summary statistics.
}
\description{
Write example GWAS summary statistics to disk.
All examples originally come from the
\href{https://www.ukbiobank.ac.uk/}{UK Biobank}.
To reduce file size, SNPs have been filtered to Minor Allele Frequency (MAF)
> 5% and a nominal p-value < 0.05. However, in practice we
 recommend using full GWAS summary statistics
  (after applying \link[MungeSumstats]{format_sumstats}).
}
\examples{
path_formatted <- MAGMA.Celltyping::get_example_gwas()
}
