% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_gwas_raw.R
\name{get_example_gwas_raw}
\alias{get_example_gwas_raw}
\title{Example GWAS: raw (pre-munged)}
\usage{
get_example_gwas_raw(
  storage_dir = tempdir(),
  trait = c("prospective_memory", "fluid_intelligence")
)
}
\arguments{
\item{save_path}{Path to save sum stats.}
}
\value{
Path to sumstats
}
\description{
Write example GWAS summary statistics to disk.
}
\keyword{internal}
