% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_example_magma_files.R
\name{get_example_magma_files}
\alias{get_example_magma_files}
\title{Get example MAGMA files}
\usage{
get_example_magma_files(file_types, storage_dir = tempdir(), verbose = TRUE)
}
\arguments{
\item{file_types}{MAGMA file types to import. 
Both ".genes.out" and ".genes.raw" are needed to run \pkg{MAGMA.Celltyping}
cell-type enrichment functions}

\item{storage_dir}{Directory to store files in.}

\item{verbose}{Print messages.}
}
\description{
Get example MAGMA files generated by 
\link[MAGMA.Celltyping]{map_snps_to_genes}.
}
\keyword{internal}
