% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gene_info.R
\name{get_gene_info}
\alias{get_gene_info}
\title{Get gene info}
\usage{
get_gene_info(genelist, attributes = c("size"), one_row_per_gene = FALSE)
}
\arguments{
\item{genelist}{Gene list.}

\item{attributes}{Attributes you want to retrieve.  A possible list of attributes can be retrieved using the function listAttributes.}
}
\description{
Get gene info from Biomart
}
\keyword{internal}
