% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_ref.R
\name{get_genome_ref}
\alias{get_genome_ref}
\title{Download genome ref}
\usage{
get_genome_ref(storage_dir = tempdir(), verbose = TRUE)
}
\arguments{
\item{storage_dir}{Where to store genome ref.}

\item{verbose}{Print messages.}
}
\value{
Directory where the genome reference data is stored.
}
\description{
Download and decompress the genome reference needed to run
\link[MAGMA.Celltyping]{map_snps_to_genes},
If the file already exists, it will simply return the path.
}
\examples{
\dontrun{
genome_ref_path <- get_genome_ref()
}
}
