% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_ref_magma.R
\name{get_genome_ref_magma}
\alias{get_genome_ref_magma}
\title{Download genome ref: magma}
\source{
\href{https://cncr.nl/research/magma/}{MAGMA site}
}
\usage{
get_genome_ref_magma(
  genome_ref_dir,
  population = "eur",
  timeout = 60 * 5,
  verbose = TRUE
)
}
\arguments{
\item{genome_ref_dir}{Folder to download reference to.}

\item{verbose}{Print messages.}
}
\description{
Download and decompress the genome reference needed to run
\link[MAGMA.Celltyping]{map_snps_to_genes},
If the file already exists, it will simply return the path.
}
\keyword{internal}
