% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_ref_piggyback.R
\name{get_genome_ref_piggyback}
\alias{get_genome_ref_piggyback}
\title{Download genome ref: piggyback}
\source{
\code{
genome_ref_path <- get_genome_ref(method = "magma")

piggyback::pb_upload(file = genome_ref_path,
                     repo = "neurogenomics/MAGMA_Celltyping",
                     overwrite = TRUE)
}
}
\usage{
get_genome_ref_piggyback(genome_ref_dir, verbose = TRUE, ...)
}
\arguments{
\item{genome_ref_dir}{Folder to download reference to.}

\item{verbose}{Print messages.}

\item{...}{
  Arguments passed on to \code{\link[=get_data]{get_data}}
  \describe{
    \item{\code{repo}}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}
    \item{\code{tag}}{tag for the GitHub release to which this data should be attached.}
    \item{\code{overwrite}}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
    \item{\code{.token}}{GitHub authentication token, see \verb{[gh::gh_token()]}}
  }}
}
\description{
Download and decompress the genome reference needed to run
\link[MAGMA.Celltyping]{map_snps_to_genes},
If the file already exists, it will simply return the path.
}
\keyword{internal}
