% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genomebuild_for_sumstats.r
\name{get_genomebuild_for_sumstats}
\alias{get_genomebuild_for_sumstats}
\title{Get genome build for sumstats}
\usage{
get_genomebuild_for_sumstats(path)
}
\arguments{
\item{path}{Filepath of the summary statistics file}
}
\value{
Either 'GRCh37' or 'GRCh38'
}
\description{
Compare the SNP:CHR:BP data against a dataset of all
Hapmap3 SNP:CHR:BP data for two genome builds (GRCh37 and GRCh38).
Evaluate which genome build matches the SNP locations from
the sum stats file.
}
\examples{
genome_build <- get_genomebuild_for_sumstats(gwas_sumstats_path)
}
