% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_magma_paths.r
\name{get_magma_paths}
\alias{get_magma_paths}
\title{Get MAGMA Paths}
\source{
\code{
gwas_sumstats_path <- MAGMA.Celltyping::get_example_gwas(
    trait = "educational_attainment")
magmaPaths <- MAGMA.Celltyping:::get_magma_paths(
    gwas_sumstats_path = gwas_sumstats_path, 
    upstream_kb = 35, 
    downstream_kb = 10)
}
}
\usage{
get_magma_paths(
  gwas_sumstats_path = NA,
  upstream_kb = NA,
  downstream_kb = NA,
  output_path = NA
)
}
\arguments{
\item{gwas_sumstats_path}{Path to the GWAS summary statistics file.}

\item{upstream_kb}{How many kb upstream of the gene 
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene 
should SNPs be included?}

\item{output_path}{If being called from 
\link[MAGMA.Celltyping]{magma_tileplot} 
then \code{gwas_sumstats_path} cannot be specified, 
so instead specificity the base folder where MAGMA_Figures exists.}
}
\value{
File path for where MAGMA files should be created
}
\description{
Gets paths to folders based on location of the GWAS summary statistics file. 
This avoids the GWAS sumstats folder becoming a mess.
}
