% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sub_SNP_LOC_DATA.R
\name{get_sub_SNP_LOC_DATA}
\alias{get_sub_SNP_LOC_DATA}
\title{Download genome ref: piggyback}
\source{
\code{ 
URL <- file.path(
    "https://github.com/neurogenomics/MAGMA_Celltyping/raw/master/",
    "data/sub_SNP_LOC_DATA.rda")
tmp <- file.path(tempdir(),basename(URL))  
piggyback::pb_upload(file = tmp,
                     repo = "neurogenomics/MAGMA_Celltyping",
                     overwrite = TRUE)
}
}
\usage{
get_sub_SNP_LOC_DATA(storage_dir = tempdir())
}
\description{
Download and decompress the genome reference needed to run
\link[MAGMA.Celltyping]{map_snps_to_genes},
If the file already exists, it will simply return the path.
}
\keyword{internal}
