% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_download_files.R
\name{github_download_files}
\alias{github_download_files}
\title{Download files from GitHub}
\usage{
github_download_files(
  filelist,
  save_dir = tempdir(),
  overwrite = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{filelist}{List of URLs to files.}

\item{save_dir}{Local directory to save files in.}

\item{overwrite}{Overwrite existing local files.}

\item{nThread}{Number of threads to parallelise downloads over.}

\item{verbose}{Print messages.}
}
\description{
Download files stored in a public GitHub repository.
}
\keyword{internal}
