% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github_list_files.R
\name{github_list_files}
\alias{github_list_files}
\title{List files}
\usage{
github_list_files(
  user = "neurogenomics",
  repo = "MAGMA_Files_Public",
  branch = c("master", "main"),
  query = NULL,
  return_download_api = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{user}{Repository user name.}

\item{repo}{Repository name.}

\item{branch}{Which branch to search.}

\item{query}{A substring search to filter files with.}

\item{return_download_api}{Return the URL to download the file 
via GitHub's API, as opposed to the URL to view the file in a web browser.}

\item{verbose}{Print messages.}
}
\value{
List of URLs.
}
\description{
List files in a GitHub repository.
}
\keyword{internal}
