% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{gwas_munged}
\alias{gwas_munged}
\title{Munged GWAS}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 93 rows and 9 columns.
}
\source{
\code{
gwas_ss <- system.file("extdata", "eduAttainOkbay.txt",
                       package = "MungeSumstats")
gwas_munged_path <- MungeSumstats::format_sumstats(gwas_ss)
gwas_munged <- data.table::fread(gwas_munged_path)
usethis::use_data(gwas_munged, overwrite = TRUE)
 }
}
\usage{
data("gwas_munged")
}
\description{
Example GWAS summary statistics standardised with
 \link[MungeSumstats]{format_sumtats}.
Original GWAS performed on Educational Attainment.
}
\keyword{datasets}
