% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{hgnc2entrez}
\alias{hgnc2entrez}
\title{HGNC to Entrez symbol mapping}
\format{
An object of class \code{data.frame} with 25839 rows and 2 columns.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
library("biomaRt")
human <- useMart(host="www.ensembl.org", "ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl")
attrib_hum = listAttributes(human)
hgnc_symbols = getBM(attributes=c("hgnc_symbol","entrezgene"), mart=human)
colnames(hgnc_symbols) = c("hgnc_symbol","entrez")
hgnc_symbols = hgnc_symbols[hgnc_symbols$hgnc_symbol!=""]
hgnc_symbols = hgnc_symbols[!is.na(hgnc_symbols$entrez),]
hgnc2entrez = hgnc_symbols
usethis::use_data(hgnc2entrez,overwrite = TRUE)
}
}
\usage{
hgnc2entrez
}
\description{
HGNC to Entrez symbol mapping
}
\keyword{datasets}
