% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_magma_files.r
\name{import_magma_files}
\alias{import_magma_files}
\title{Import preprocessed MAGMA files}
\source{
\href{https://github.com/neurogenomics/MAGMA_Files_Public}{
\emph{MAGMA_Files_Public} GitHub repository}
}
\usage{
import_magma_files(
  save_dir = tempdir(),
  ids = NULL,
  file_types = c(".genes.out", ".genes.raw"),
  overwrite = FALSE,
  return_dir = TRUE,
  return_tables = FALSE,
  nested = FALSE,
  nThread = 1,
  verbose = TRUE
)
}
\arguments{
\item{save_dir}{Parent folder where you want to save the MAGMA files.}

\item{ids}{[Optional] \href{https://gwas.mrcieu.ac.uk/}{OpenGWAS} dataset IDs. 
You see metadata for the GWAS 
\href{https://github.com/neurogenomics/MAGMA_Files_Public/blob/master/metadata.csv}{here}.}

\item{file_types}{MAGMA file types to import. 
Both ".genes.out" and ".genes.raw" are needed to run \pkg{MAGMA.Celltyping}
cell-type enrichment functions}

\item{overwrite}{If the files have already been downloaded in 
the specified directory,
these downloads will be skipped. Set \code{overwrite=TRUE} to force 
them to be re-downloaded (Default: \code{FALSE}).}

\item{return_dir}{Return a list of unique directory names instead of the 
full file paths (Default: \code{TRUE}).}

\item{return_tables}{Returns the files imported as tables (\code{TRUE}),
instead of as paths (default: \code{FALSE}). 
Overrides \code{return_dir} argument.}

\item{nested}{Return files as a list nested by \code{id} (default: \code{FALSE}).}

\item{nThread}{Number of threads to parallelise downloads across.}

\item{verbose}{Print messages.}
}
\value{
Named vector of paths to downloaded MAGMA files (or directories), 
or a \link[data.table]{data.table}.
}
\description{
Import preprocessed MAGMA SNPs-to-genes mapping files for many GWAS. 
These files were generated using 
\link[MAGMA.Celltyping]{map_snps_to_genes} and are publicly available via the 
\href{https://github.com/neurogenomics/MAGMA_Files_Public}{
\emph{MAGMA_Files_Public} GitHub repository}. 
Metadata for each dataset (including trait descriptions) can be found here 
\href{https://github.com/neurogenomics/MAGMA_Files_Public/blob/master/metadata.csv}{here}.
}
\examples{
magma_dirs <- MAGMA.Celltyping::import_magma_files(ids = c("ieu-a-298"))
}
