% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_magma_files_metadata.R
\name{import_magma_files_metadata}
\alias{import_magma_files_metadata}
\title{Import MAGMA files metadata}
\usage{
import_magma_files_metadata(file_types, use_local = TRUE, verbose = TRUE)
}
\arguments{
\item{file_types}{MAGMA file types to import. 
Both ".genes.out" and ".genes.raw" are needed to run \pkg{MAGMA.Celltyping}
cell-type enrichment functions}

\item{use_local}{Whether to use a local copy of the metadata.
\itemize{
\item{\code{TRUE}}{Use the metadata included with \pkg{MAGMA.Celltyping}
 (less frequently updated).}  
\item{\code{FALSE}}{Import metadata from the GitHub repository 
(more frequently updated).}  
}}

\item{verbose}{Print messages.}
}
\description{
Import metadata from the SNP-to-gene mapping files available in 
\href{https://github.com/neurogenomics/MAGMA_Files_Public}{
MAMGA_Files_Public}.
}
\keyword{internal}
