% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_magma_files_tables.R
\name{import_magma_files_tables}
\alias{import_magma_files_tables}
\title{Import magma files as tables}
\usage{
import_magma_files_tables(local_files, verbose = TRUE)
}
\arguments{
\item{local_files}{A list of magma files, nested by GWAS ID.}

\item{verbose}{Print messages.}
}
\description{
Import magma \emph{genes.out} and \emph{genes.raw} 
output files as a nested list of \link[data.table]{data.table}s.
}
\keyword{internal}
