% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infer_geneset_species.R
\name{infer_geneset_species}
\alias{infer_geneset_species}
\title{Infer gene set species}
\usage{
infer_geneset_species(geneset, geneset_species = NULL, verbose = 1, ...)
}
\arguments{
\item{geneset}{Genes which are to be tested. 
Can be gene symbols (or other gene IDs) from any species listed in 
\link[EWCE]{list_species}.}

\item{geneset_species}{Species that the \code{geneset} came from.
Can be any species listed in \link[EWCE]{list_species}.
If \code{geneset_species!="human"}, the \code{geneset} will be converted
 to 1:1 human orthologs using \link[orthogene]{convert_orthologs}.}

\item{verbose}{Message verbosity. 
\itemize{
\item{\code{0} or \code{FALSE} : }{
Don't print any messages.
}
\item{\code{1} or \code{TRUE} : }{
Only print messages from \pkg{MAGMA.Celltyping}.
}
\item{\code{2} or \code{c(TRUE,TRUE)} : }{
Print messages from \pkg{MAGMA.Celltyping} and 
the internal \pkg{orthogene} function.}
}}

\item{...}{
  Arguments passed on to \code{\link[orthogene:infer_species]{orthogene::infer_species}}
  \describe{
    \item{\code{gene_df}}{Data object containing the genes
(see \code{gene_input} for options on how
the genes can be stored within the object).\cr
Can be one of the following formats:\cr
\itemize{
\item{\code{matrix} : \cr}{A sparse or dense matrix.}
\item{\code{data.frame} : \cr}{A \code{data.frame},
 \code{data.table}. or \code{tibble}.}
\item{code{list} : \cr}{A \code{list} or character \code{vector}.}
}
Genes, transcripts, proteins, SNPs, or genomic ranges
 can be provided in any format
(HGNC, Ensembl, RefSeq, UniProt, etc.) and will be
automatically converted to gene symbols unless
specified otherwise with the \code{...} arguments.\cr
\emph{Note}: If you set \code{method="homologene"}, you
must either supply genes in gene symbol format (e.g. "Sox2")
 OR set \code{standardise_genes=TRUE}.}
    \item{\code{gene_input}}{Which aspect of \code{gene_df} to
get gene names from:\cr
\itemize{
\item{\code{"rownames"} : \cr}{From row names of data.frame/matrix.}
\item{\code{"colnames"} : \cr}{From column names of data.frame/matrix.}
\item{\code{<column name>} : \cr}{From a column in \code{gene_df},
 e.g. \code{"gene_names"}.}
}}
    \item{\code{test_species}}{Which species to test for matches with.
If set to \code{NULL}, will default to a list of humans and
 5 common model organisms.
 If \code{test_species} is set to one of the following options, 
 it will automatically pull all species from that respective package and 
 test against each of them:  
 \itemize{
 \item{"homologene" : }{20+ species (default)}
 \item{"gprofiler" : }{700+ species}
 \item{"babelgene" : }{19 species}
 }}
    \item{\code{method}}{R package to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives consensus scores for each gene mapping based on a
 several different data sources.}
}}
    \item{\code{make_plot}}{Make a plot of the results.}
    \item{\code{show_plot}}{Print the plot of the results.}
  }}
}
\value{
Inferred species name.
}
\description{
Infers species from from a gene set
using \link[orthogene]{infer_species}. 
If \code{geneset_species} is not \code{NULL}, 
this will be returned instead of inferring the species.
}
\examples{
  
geneset <- MAGMA.Celltyping::rbfox_binding
geneset_species <- infer_geneset_species(geneset = geneset)
}
