% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_magma.R
\name{install_magma}
\alias{install_magma}
\title{Install the MAGMA command line tool}
\source{
\href{https://ctg.cncr.nl/software/magma}{MAGMA website}

\href{https://github.com/NathanSkene/MAGMA_Celltyping}{
MAGMA.celltyping documentation}
}
\usage{
install_magma(dest_dir = "/usr/local/bin", upgrade = FALSE, verbose = TRUE)
}
\arguments{
\item{dest_dir}{Folder in which to install MAGMA.}

\item{verbose}{Print messages.}

\item{version}{Which version of MAGMA to install.}
}
\description{
Install the MAGMA command line tool
}
