% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/invert_dict.R
\name{invert_dict}
\alias{invert_dict}
\title{Invert dictionary}
\usage{
invert_dict(dict)
}
\arguments{
\item{dict}{Named vector.}
}
\description{
Swap the names/values from a named vector.
}
\keyword{internal}
