% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_gsea.R
\name{iterate_gsea}
\alias{iterate_gsea}
\title{Iterate gene set enrichment analysis across all traits and celltypes}
\usage{
iterate_gsea(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  x_quantiles = 10,
  y_quantiles = 10,
  use_quantiles = 10,
  nCores = 1
)
}
\arguments{
\item{xmat}{gene x trait matrix.}

\item{ymat}{gene x celltype matrix.}

\item{correction_method}{Multiple-testing correction method to be passed to \code{stats::p.adjust}.}

\item{qvalue_thresh}{q.value threshold to use when report significant results summary.}

\item{x_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{y_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{nCores}{Number of cores to use in parallel. Will optimize if \code{NULL}.}
}
\description{
Iterate gene set enrichment analysis across all traits and celltypes
}
