% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/iterate_lm.R
\name{iterate_lm}
\alias{iterate_lm}
\title{Iterate linear regression across all traits and celltypes}
\usage{
iterate_lm(
  xmat,
  ymat,
  correction_method = "BH",
  qvalue_thresh = 0.05,
  y_quantiles = NULL,
  nCores = NULL
)
}
\arguments{
\item{xmat}{gene x trait matrix.}

\item{ymat}{gene x celltype matrix.}

\item{correction_method}{Multiple-testing correction method to be passed to \code{stats::p.adjust}.}

\item{qvalue_thresh}{q.value threshold to use when report significant results summary.}

\item{y_quantiles}{The number of quantiles to bin \code{ymat} data into.}

\item{nCores}{Number of cores to use in parallel. Will optimize if \code{NULL}.}
}
\description{
Iterate linear regression across all traits and celltypes
}
