% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_remote_files.R
\name{list_remote_files}
\alias{list_remote_files}
\title{List files on a remote server or website}
\source{
\href{https://stackoverflow.com/questions/15954463/read-list-of-file-names-from-web-into-r}{
stackoverflow}
}
\usage{
list_remote_files(URL, pattern = "*.zip")
}
\description{
List files on a remote server or website
}
