% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.magma.results.file.r
\name{load.magma.results.file}
\alias{load.magma.results.file}
\title{Load a MAGMA .gcov.out file}
\usage{
load.magma.results.file(path, annotLevel, ctd)
}
\arguments{
\item{path}{Path to a .gcov.out file}

\item{annotLevel}{Which annotation level does this .gcov.out file relate to?}

\item{ctd}{Cell type data strucutre containing $quantiles}
}
\value{
Contents of the .gcov.out file
}
\description{
Convenience function which just does a little formatting to make it easier to use
}
\examples{
res = load.magma.results.file(path="Raw/adhd_formtcojo.txt.NEW.10UP.1DOWN.gov.out",annotLevel=1,ctd=ctd)

}
