% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_magma_results_file.r
\name{load_magma_results_file}
\alias{load_magma_results_file}
\title{Load a MAGMA .gcov.out file}
\usage{
load_magma_results_file(
  path,
  annotLevel,
  ctd,
  genesOutCOND = NA,
  EnrichmentMode = "Linear",
  ControlForCT = "BASELINE"
)
}
\arguments{
\item{path}{Path to a .gcov.out file (if \code{EnrichmentMode=='Linear'}).
or .sets.out (if EnrichmentMode=='Top 10\%').}

\item{annotLevel}{Level within the \code{ctd} to use.}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{genesOutCOND}{[Optional] If the analysis controlled for another GWAS, 
then this is included as a column. Otherwise the column is \code{NA}.}

\item{EnrichmentMode}{[Optional] Either 'Linear' or 'Top 10\%' 
(Default: \code{'Linear'}).}

\item{ControlForCT}{[Optional] May be an internal argument. 
We suggest ignoring or take a look at the code to figure it out.}
}
\value{
Contents of the .gcov.out file.
}
\description{
Convenience function which just does a little formatting to make 
it easier to use.
}
\keyword{internal}
