% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_rdata.R
\name{load_rdata}
\alias{load_rdata}
\title{Load RData}
\usage{
load_rdata(fileName)
}
\arguments{
\item{fileName}{Path to .rda file.}
}
\description{
Load a .rda file directly into a specific variable, 
This avoids the issue of not knowing what the variable name will be when 
loadings .rda files the usual way (\code{load}).
}
\keyword{internal}
