% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma.tileplot.r
\name{magma.tileplot}
\alias{magma.tileplot}
\title{MAGMA tileplot}
\usage{
magma.tileplot(ctd, results, height = 13, width = 4, annotLevel = 1,
  fileTag = "", output_path)
}
\arguments{
\item{ctd}{Cell type data strucutre containing $quantiles}

\item{height}{Height of the output tileplot}

\item{width}{Width of the output tileplot}

\item{annotLevel}{Annotation level to plot the results for}

\item{output_path}{Location where the results should be plotted}

\item{ctAssoc1}{Output from either merge_magma_results()}
}
\description{
Used after merging results from multiple GWAS studies
}
\examples{
ctAssocs = merge_magma_results.r(ctAssoc1,ctAssoc2)

}
