% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_check.R
\name{magma_check}
\alias{magma_check}
\title{MAGMA check}
\usage{
magma_check(
  version = NULL,
  upgrade = FALSE,
  return_version = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{version}{MAGMA version.}

\item{upgrade}{If MAGMA is already installed and is not the latest version, 
should it be upgraded to the latest version? (Default: \code{FALSE}).}

\item{return_version}{Return the MAGMA version being used.}

\item{verbose}{Print messages.}
}
\description{
Check that MAGMA is installed at the beginning of functions where MAGMA is 
required.
}
\keyword{internal}
