% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_check_version_match.R
\name{magma_check_version_match}
\alias{magma_check_version_match}
\title{MAGMA check version match}
\usage{
magma_check_version_match(desired_version, verbose = TRUE)
}
\arguments{
\item{desired_version}{Desired version of MAGMA.}

\item{verbose}{Print messages.}
}
\description{
Check that the desired version of MAGMA is installed.
}
\keyword{internal}
