% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_download_binary.R
\name{magma_download_binary}
\alias{magma_download_binary}
\title{Download a MAGMA binary}
\usage{
magma_download_binary(
  magma_url,
  dest_dir = find_install_dir(),
  destfile = basename(magma_url),
  verbose = TRUE
)
}
\arguments{
\item{magma_url}{URL path to MAGMA executable.}

\item{dest_dir}{Local directory to store the MAGMA executable in.}

\item{verbose}{Print messages.}
}
\description{
Download a MAGMA binary executable file.
}
\keyword{internal}
