% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_executable.R
\name{magma_executable}
\alias{magma_executable}
\title{Find the local executable}
\source{
\href{https://ctg.cncr.nl/software/MAGMA/prog/archive/}{
MAGMA archive}
}
\usage{
magma_executable(
  exec_dir = find_install_dir(verbose = FALSE),
  check_exists = FALSE,
  version = NULL,
  verbose = FALSE
)
}
\arguments{
\item{verbose}{Print messages.}
}
\value{
Path to appropriate executable
}
\description{
Find the local executable for a specific OS.
}
\keyword{internal}
