% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_executable_select.R
\name{magma_executable_select}
\alias{magma_executable_select}
\title{Select a MAGMA executable}
\usage{
magma_executable_select(magma_x, return_all = FALSE, verbose = TRUE)
}
\arguments{
\item{magma_x}{MAGMA executable.}

\item{return_all}{Whether to return paths to all MAGMA executables.}

\item{verbose}{Print messages.}
}
\description{
When more than one MAGMA executable is installed, 
selects one of them to use.
}
\keyword{internal}
