% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_find_executable.R
\name{magma_find_executable}
\alias{magma_find_executable}
\title{Find MAGMA executable}
\usage{
magma_find_executable(destpath, return_all = TRUE)
}
\arguments{
\item{destpath}{Directory to search for MAGMA executables in.}

\item{return_all}{Whether to return paths to all MAGMA executables.}
}
\description{
Find any existing MAGMA executables in a given directory.
}
\keyword{internal}
