% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_installed.R
\name{magma_installed}
\alias{magma_installed}
\title{MAGMA installed}
\usage{
magma_installed(magma_x = magma_executable(), verbose = TRUE)
}
\arguments{
\item{magma_x}{MAGMA executable.}

\item{verbose}{Print messages.}
}
\description{
Test whether any version of MAGMA is installed.
}
\keyword{internal}
