% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_installed_version.R
\name{magma_installed_version}
\alias{magma_installed_version}
\title{MAGMA installed version}
\usage{
magma_installed_version(magma_x = magma_executable(), verbose = TRUE)
}
\arguments{
\item{magma_x}{MAGMA executable.}

\item{verbose}{Print messages.}
}
\description{
List which version(s) of MAGMA is currently installed.
}
\keyword{internal}
