% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_links.R
\name{magma_links}
\alias{magma_links}
\title{Get MAGMA links}
\usage{
magma_links(
  latest_only = TRUE,
  os = get_os(),
  version = NULL,
  use_local = TRUE,
  return_table = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{latest_only}{Only return the latest version of MAGMA.}

\item{os}{Operating System.}

\item{version}{Desired MAGMA version.}

\item{use_local}{Use a copy of the MAGMA file links, which may be out of date
but will always be available.}

\item{return_table}{Return the results in table format.}

\item{verbose}{Print messages.}
}
\description{
Get links to MAGMA executables in the official MAGMA archives.
}
\keyword{internal}
