% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_links_gather.R
\name{magma_links_gather}
\alias{magma_links_gather}
\title{Gather MAGMA links}
\usage{
magma_links_gather()
}
\description{
Gather links to MAGMA executables stored in the official MAGMA archives, 
and format them as a table.
}
\keyword{internal}
