% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_links_query.R
\name{magma_links_query}
\alias{magma_links_query}
\title{Query MAGMA links}
\usage{
magma_links_query(latest_only)
}
\arguments{
\item{latest_only}{Only return the latest version of MAGMA.}
}
\description{
Gather links to MAGMA executables stored in the official MAGMA archives.
}
\keyword{internal}
