% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{magma_links_stored}
\alias{magma_links_stored}
\title{Stored links to MAGMA executables}
\format{
An object of class \code{data.table} (inherits from \code{data.frame}) with 121 rows and 5 columns.
}
\source{
\href{https://ctg.cncr.nl/software/magma}{MAGMA}

\code{
magma_links_stored <- MAGMA.Celltyping:::magma_links_gather()
usethis::use_data(magma_links_stored, overwrite = TRUE)
}
}
\usage{
data("magma_links_stored")
}
\description{
Stored links to MAGMA executables in the 
\href{https://ctg.cncr.nl/software/MAGMA/prog/archive/}{MAGMA archives}.
}
\keyword{datasets}
