% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_links_versions.R
\name{magma_links_versions}
\alias{magma_links_versions}
\title{Get the available versions of MAGMA}
\source{
\href{https://ctg.cncr.nl/software/magma}{MAGMA website}
}
\usage{
magma_links_versions(links, version = NULL, return_all = FALSE, verbose = TRUE)
}
\value{
Character vector of versions.
}
\description{
Get the available versions of MAGMA
}
\keyword{internal}
