% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_os_suffix.R
\name{magma_os_suffix}
\alias{magma_os_suffix}
\title{Get MAGMA OS suffix}
\usage{
magma_os_suffix(os = get_os())
}
\arguments{
\item{os}{Operating System.}
}
\description{
Get the corresponding MAGMA executable file suffix for a given OS.
}
\keyword{internal}
