% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_read_covar_out.R
\name{magma_read_covar_out}
\alias{magma_read_covar_out}
\title{MAGMA: read gsa.out}
\usage{
magma_read_covar_out(out_prefix)
}
\arguments{
\item{out_prefix}{Out file path prefix.}
}
\description{
Read a gsa.out file generated by MAGMA enrichment tests.
}
\keyword{internal}
