% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_read_gsa_out.R
\name{magma_read_gsa_out}
\alias{magma_read_gsa_out}
\title{MAGMA: read gsa.out}
\usage{
magma_read_gsa_out(out_prefix, analysis_name)
}
\arguments{
\item{out_prefix}{Out file path prefix.}

\item{analysis_name}{Used in file names which area created.}
}
\description{
Read a gsa.out file generated by MAGMA enrichment tests.
}
\keyword{internal}
