% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_read_sets_out.R
\name{magma_read_sets_out}
\alias{magma_read_sets_out}
\title{MAGMA: read sets.out}
\usage{
magma_read_sets_out(out_prefix, verbose = TRUE)
}
\arguments{
\item{out_prefix}{Out file path prefix.}

\item{verbose}{Print messages.}
}
\description{
Read a sets.out file generated by MAGMA enrichment tests.
}
\keyword{internal}
