% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_run.R
\name{magma_run}
\alias{magma_run}
\title{Run MAGMA}
\source{
\href{https://ctg.cncr.nl/software/MAGMA/prog/archive/}{
MAGMA archive}
}
\usage{
magma_run(version = NULL, cmd = NULL, verbose = TRUE)
}
\arguments{
\item{version}{MAGMA version to use.}

\item{cmd}{Commands to MAGMA as a string or list of strings.}

\item{verbose}{Print messages.}
}
\description{
Find the local MAGMA executable for a specific OS and run a command with it.
}
\examples{
MAGMA.Celltyping::magma_run(cmd = "--version")
}
