% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_uninstall.R
\name{magma_uninstall}
\alias{magma_uninstall}
\title{Uninstall MAGMA}
\usage{
magma_uninstall(version = NULL, verbose = TRUE)
}
\arguments{
\item{version}{Which version of MAGMA to uninstall. If \code{NULL}, 
will uninstall all versions found.}

\item{verbose}{Print messages.}
}
\value{
Path(s) to MAGMA installation folder(s) that were deleted.
}
\description{
Uninstall one or all versions of MAGMA.
}
\examples{
MAGMA.Celltyping::magma_uninstall()
}
