% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_versions.R
\name{magma_versions}
\alias{magma_versions}
\title{Get the available versions of MAGMA}
\source{
\href{https://ctg.cncr.nl/software/magma}{MAGMA website}
}
\usage{
magma_versions(links)
}
\description{
Get the available versions of MAGMA
}
\keyword{internal}
