% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mungesumstats_deprecation_msg.R
\name{mungesumstats_deprecation_msg}
\alias{mungesumstats_deprecation_msg}
\title{Deprecation notice}
\usage{
mungesumstats_deprecation_msg(error = TRUE)
}
\arguments{
\item{error}{Whether to return an error or simply a message.}
}
\description{
Deprecation notice for functions now performed by \pkg{MungeSumstats}.
}
\keyword{internal}
