% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_associations.r
\name{plot_celltype_associations}
\alias{plot_celltype_associations}
\title{Plot celltype associations calculated using MAGMA}
\usage{
plot_celltype_associations(
  ctAssocs,
  ctd,
  useSignificanceLine = TRUE,
  savePDF = TRUE,
  fileTag = "",
  plotDendro = TRUE,
  gwas_title = "",
  plotLegend = TRUE,
  figsDir = NULL,
  show_plot = TRUE,
  bind_plots = FALSE,
  ncol = 1,
  verbose = TRUE
)
}
\arguments{
\item{ctAssocs}{Output from either 
\link[MAGMA.Celltyping]{calculate_celltype_associations} or
\link[MAGMA.Celltyping]{calculate_conditional_celltype_associations}.}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{useSignificanceLine}{TRUE or FALSE. Should their be a vertical
line marking Bonferroni significance?}

\item{savePDF}{TRUE or FALSE. Save figure to file or print to screen?}

\item{fileTag}{String appended to the names of the saved PDFs,
i.e. the name of the celltype data file used}

\item{plotDendro}{Should the dendrogram of celltypes be
shown alongside the figure? TRUE or FALSE.}

\item{gwas_title}{Title to be displayed over the figure (string).}

\item{plotLegend}{Should the figure legend be displayed?}

\item{figsDir}{Directory where figures should be created.}

\item{show_plot}{Print to the plot(s).}

\item{bind_plots}{Bind the list of plots together using 
\link[patchwork]{wrap_plots}.}

\item{ncol}{Number of columns to put plots into during \code{bind_plots}.}

\item{verbose}{Print messages.}
}
\value{
A list of ggplot objects.
}
\description{
Can take input from either
 \link[MAGMA.Celltyping]{calculate_celltype_associations} or
 \link[MAGMA.Celltyping]{calculate_conditional_celltype_associations}.
}
\examples{
ctAssocs <- MAGMA.Celltyping::enrichment_results[[1]]$ctAssocsLinear
ctd <- ewceData::ctd()
figs <- plot_celltype_associations(ctAssocs = ctAssocs,
                                   ctd = ctd,
                                   savePDF = FALSE)
}
