% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prepare_quantile_groups.r
\name{prepare_quantile_groups}
\alias{prepare_quantile_groups}
\title{Prepare quantile groups for each celltype based on specificity}
\usage{
prepare_quantile_groups(
  ctd,
  specificity_species = "mouse",
  gwas_species = "human",
  bins = 40
)
}
\arguments{
\item{ctd}{Cell type data}

\item{specificity_species}{Species name relevant to the cell type data,
(e.g. "mouse" or "human").}

\item{gwas_species}{Species name relevant to the GWAS data, in almost all
cases this will be "human"}

\item{bins}{How many bins should specificity be divided into?}
}
\value{
The ctd with additional quantiles matrix
}
\description{
Quantile groups are stored in an extra matrix ('quantiles')
in the returned CTD. This function also removes any genes
from the CTD data which are not 1:1 orthologs with the GWAS species.
}
\examples{
\dontrun{
ctd <- prepare_quantile_groups(ctd = ewceData::ctd())
}

}
