% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_magma_genes_out.R
\name{read_magma_genes_out}
\alias{read_magma_genes_out}
\title{Read a \emph{.genes.out} file}
\usage{
read_magma_genes_out(path, verbose = TRUE)
}
\arguments{
\item{path}{Path to a MAGMA \emph{.genes.out} file.}

\item{verbose}{Print messages.}
}
\value{
\link[data.table]{data.table}
}
\description{
Read a \emph{.genes.out} file  generated by
 \link[MAGMA.Celltyping]{map_snps_to_genes} 
 as a \link[data.table]{data.table}.
}
\keyword{internal}
