% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regress_gene_info.R
\name{regress_gene_info}
\alias{regress_gene_info}
\title{Regress out the effect of gene attributes}
\usage{
regress_gene_info(xmat, attributes = c("size"), correction_method = "BH")
}
\arguments{
\item{xmat}{gene x sample matrix.}

\item{attributes}{Gene attributes to extract from \code{biomaRt}
and then regress from \code{xmat}.}

\item{correction_method}{Multiple-testing correction method to be passed to \code{stats::p.adjust}.}
}
\description{
Queries \code{biomaRt} for \code{gene_attributes} and then regresses them out of \code{xmat}.
}
\examples{
library(MAGMA.Celltyping)
xmat <- MAGMA.Celltyping::ctd_DRONC_human[[1]]$mean_exp[1:100,]
adjusted_df <- regress_gene_info(xmat) 
}
