% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_barplot.R
\name{results_barplot}
\alias{results_barplot}
\title{Plot barplot of enrichment results - plotting -log10 (p-value) but cut-off
significance is based on FDR where if either linear or top10% result is <
FDR threshold the result will be plotted}
\usage{
results_barplot(
  merged_results,
  title = NULL,
  y_lab = NULL,
  x_lab = "-log10(p-value)",
  fdr_thresh = 0.05,
  y_var = "Celltype",
  x_var = "-log10p",
  fill_var = "EnrichmentMode",
  fill_cols = c("#60A9BD", "#3A62A0"),
  horz_line_p = 0.01,
  sort_celltypes_by = "Top 10\%",
  show_plot = TRUE,
  height = 5,
  width = 7,
  dpi = 300,
  save_path = file.path(tempdir(), "MAGMA_Celltyping.barplot.jpg")
)
}
\arguments{
\item{merged_results}{Enrichment results generated by 
\link[MAGMA.Celltyping]{celltype_associations_pipeline} and merged by 
\link[MAGMA.Celltyping]{merge_results}.}

\item{title}{Plot title.}

\item{y_lab}{Plot y-axis label.}

\item{x_lab}{Plot x-axis label.}

\item{fdr_thresh}{FDR filtering threshold.}

\item{y_var}{y-axis variable.}

\item{x_var}{x-axis variable.}

\item{fill_var}{Fill variable.}

\item{fill_cols}{The two colours to use for the fill param (Vector).}

\item{horz_line_p}{add dashed line for a p-value cut-off - default of 0.01}

\item{sort_celltypes_by}{Sort the order of the cell types in the plot by the
enrichment results. Choose to sort by Top 10\% or Linear results or set to
NULL to avoid sorting. Default is Top 10\%.}

\item{show_plot}{Whether to print the plot.}

\item{height}{Plot height.}

\item{width}{Plot width.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{save_path}{Path to save plot to.}
}
\value{
ggplot object.
}
\description{
Plot of cell-type enrichment results from 
\link[MAGMA.Celltyping]{celltype_associations_pipeline}.
}
\examples{
MAGMA_results <- MAGMA.Celltyping::enrichment_results
merged_results <- MAGMA.Celltyping::merge_results(MAGMA_results)
bar <- MAGMA.Celltyping::results_barplot(
    merged_results = merged_results,
    fdr_thresh = 1)  
}
