% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_heatmap.R
\name{results_heatmap}
\alias{results_heatmap}
\title{Plot heatmap of GWAS x cell-type enrichment results}
\usage{
results_heatmap(
  merged_results,
  title = NULL,
  x_lab = NULL,
  fdr_thresh = 0.05,
  facet_formula = "EnrichmentMode ~ .",
  x_var = "Celltype",
  y_var = "GWAS",
  fill_var = "-log1p(FDR)",
  scales = "free_y",
  space = "fixed",
  show_plot = TRUE,
  height = 5,
  width = 7,
  dpi = 300,
  save_path = file.path(tempdir(), "MAGMA_Celltyping.heatmap.jpg")
)
}
\description{
Plot enrichment results from 
\link[MAGMA.Celltyping]{celltype_associations_pipeline}.
}
\examples{
MAGMA_results <- MAGMA.Celltyping::enrichment_results
merged_results <- MAGMA.Celltyping::merge_results(res)
heat <- MAGMA.Celltyping::results_heatmap(
    merged_results = merged_results,
    fdr_thresh = 1)  
}
