% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set_permissions.R
\name{set_permissions}
\alias{set_permissions}
\title{Set permissions}
\usage{
set_permissions(path, is_folder = FALSE, verbose = TRUE)
}
\arguments{
\item{path}{Path to file/folder.}

\item{is_folder}{Whether \code{path} is a folder. 
If so, permissions will be set recursively for all files/subfolders.}

\item{verbose}{Print messages.}
}
\description{
Robust methods for setting file/folder permissions across multiple OS.
}
\keyword{internal}
