% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{SNP_LOC_DATA}
\alias{SNP_LOC_DATA}
\title{RSD / CHR / BP for GRCh37 and GRCh38}
\format{An object of class \code{data.frame} with 2428938 rows and 4 columns.}
\source{
build_snp_location_tables()
}
\usage{
SNP_LOC_DATA
}
\description{
Data table with all Hapmap3 SNPs
}
\keyword{datasets}
