% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.sumstats.column.headers.R
\name{standardise.sumstats.column.headers}
\alias{standardise.sumstats.column.headers}
\title{Standardise the column headers in the Summary Statistics files}
\usage{
standardise.sumstats.column.headers(path)
}
\arguments{
\item{path}{File path for the summary statistics file.}
}
\value{
The amended column headers 
(also the column headers will be written directly 
into the summary statistics file)
}
\description{
Use a reference data table of common column header names
 (stored as \link[MAGMA.Celltyping]{sumstatsColHeaders}) 
 convert them to a standard set, i.e. chromosome --> CHR.
This function does not check that all the required column headers
 are present.
The amended header is written directly back into the file.
}
\examples{
\dontrun{ 
path <- MAGMA.Celltyping::get_example_gwas()
col_headers <- MAGMA.Celltyping::standardise.sumstats.column.headers(
    path = path)
}
}
