% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sub_SNP_LOC_DATA}
\alias{sub_SNP_LOC_DATA}
\title{Subsampled SNP location data}
\format{
An object of class \code{data.frame} with 100000 rows and 4 columns.
}
\source{
\dontrun{
 filePath = sprintf("%s/MAGMA.Celltyping/data/SNP_LOC_DATA.rda",.libPaths()[1])
 load(filePath)
 sub_SNP_LOC_DATA = SNP_LOC_DATA[sample(1:dim(SNP_LOC_DATA)[1],100000),]
 usethis::use_data(sub_SNP_LOC_DATA, overwrite = TRUE)
}
}
\usage{
sub_SNP_LOC_DATA
}
\description{
SNP location data used for inferring genome build 
in \code{get_genomebuild_for_sumstats}.
}
\keyword{datasets}
