% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{sumstatsColHeaders}
\alias{sumstatsColHeaders}
\title{Summary Statistics Column Headers}
\format{
An object of class \code{data.frame} with 82 rows and 2 columns.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
# Most the data in the below table comes from the LDSC github wiki
sumstatsColHeaders = read.csv("Magma_Column_headers.csv",stringsAsFactors = FALSE)
usethis::use_data(sumstatsColHeaders,overwrite = TRUE)
}
}
\usage{
sumstatsColHeaders
}
\description{
List of uncorrected column headers often found in GWAS Summary Statistics column headers
}
\keyword{datasets}
