% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust.zstat.in.genesOut.r
\name{adjust.zstat.in.genesOut}
\alias{adjust.zstat.in.genesOut}
\title{Adjust MAGMA Z-statistic from .genes.out files}
\usage{
adjust.zstat.in.genesOut(ctd, magma_GenesOut_file = NA, sctSpecies = "mouse")
}
\arguments{
\item{ctd}{Cell type data structure}

\item{magma_GenesOut_file}{A MAGMA .genes.out file}

\item{sctSpecies}{Either 'human' or 'mouse'}
}
\description{
Used when you want to directly analyse the gene level z-scores corrected for gene length etc
}
\examples{
myGenesOut = tempfile()
data.table::fwrite(x=MAGMA.Celltyping::genesOut,sep="\t",file=myGenesOut)
magmaGenesOut = adjust.zstat.in.genesOut(ewceData::ctd(),
magma_GenesOut_file=myGenesOut,sctSpecies="mouse")

}
