% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_zstat_in_genesOut.r
\name{adjust_zstat_in_genesOut}
\alias{adjust_zstat_in_genesOut}
\title{Adjust MAGMA Z-statistic from .genes.out files}
\usage{
adjust_zstat_in_genesOut(
  ctd,
  magma_GenesOut_file = NA,
  ctd_species = "mouse",
  output_species = "human",
  method = "bonferroni",
  ...
)
}
\arguments{
\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{magma_GenesOut_file}{A MAGMA .genes.out file.}

\item{ctd_species}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}

\item{output_species}{Which species' genes names to convert \code{exp} to.
See \link[EWCE]{list_species} for all available species.}

\item{method}{R package to to use for gene mapping:
\itemize{
\item{\code{"gprofiler"} : Slower but more species and genes.}
\item{\code{"homologene"} : Faster but fewer species and genes.}
\item{\code{"babelgene"} : Faster but fewer species and genes.
Also gives }
}
 (slower but more species and genes) or
\code{"homologene"} (faster but fewer species and genes).}

\item{...}{Additional arguments passed for \link[EWCE]{standardise_ctd}.}
}
\description{
Used when you want to directly analyse the gene level
 z-scores corrected for gene length etc
}
\examples{
myGenesOut <- MAGMA.Celltyping::import_magma_files(
    ids = c("ieu-a-298"),
    file_types = ".genes.out",
    return_dir = FALSE)
ctd <- ewceData::ctd()

magmaGenesOut <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd,
    magma_GenesOut_file = myGenesOut,
    ctd_species = "mouse"
)
}
