% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_zstat_in_genesOut.r
\name{adjust_zstat_in_genesOut}
\alias{adjust_zstat_in_genesOut}
\title{Adjust MAGMA Z-statistic from .genes.out files}
\usage{
adjust_zstat_in_genesOut(ctd, magma_GenesOut_file = NA, sctSpecies = "mouse")
}
\arguments{
\item{ctd}{Cell type data structure}

\item{magma_GenesOut_file}{A MAGMA .genes.out file}

\item{sctSpecies}{Either 'human' or 'mouse'}
}
\description{
Used when you want to directly analyse the gene level
 z-scores corrected for gene length etc
}
\examples{
myGenesOut <- tempfile()
data.table::fwrite(
    x = MAGMA.Celltyping::genesOut,
    file = myGenesOut,
    sep = "\t"
)
magmaGenesOut <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ewceData::ctd(),
    magma_GenesOut_file = myGenesOut,
    sctSpecies = "mouse"
)
}
