% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{all_hgnc_wtEntrez}
\alias{all_hgnc_wtEntrez}
\title{All HGNC gene symbols with ENTREZ gene IDs}
\format{
An object of class \code{data.frame} with 60617 rows and 2 columns.
}
\source{
The code to prepare the .Rda file file from the marker file is:
\code{
#library("biomaRt")
#listMarts(host="www.ensembl.org")
#human <- useMart(host="www.ensembl.org",
#"ENSEMBL_MART_ENSEMBL", dataset="hsapiens_gene_ensembl")
#all_hgnc_wtEntrez = getBM(attributes=c("hgnc_symbol","entrezgene"),
#mart=human)
download.file("ftp://ftp.ncbi.nlm.nih.gov/gene/DATA/gene_info.gz")
# Note, 9606 is the identifier for human, use 10090 for mouse
system(paste("gzip -cd gene_info.gz | awk \'BEGIN {FS=\"\t\"} $1==9606",
             "{print $2 \"\t\" $3 \"\t\" $10}\' > geneInfo.txt)"))

allHGNC=read.table(file="/Users/natske/geneInfo.txt",fill  = TRUE)
all_hgnc_wtEntrez = allHGNC[,seq(2,1)]
colnames(all_hgnc_wtEntrez) = c("hgnc_symbol","entrezgene")
all_hgnc_wtEntrez = unique(all_hgnc_wtEntrez)
usethis::use_data(all_hgnc_wtEntrez,overwrite = TRUE)
}
}
\usage{
data("all_hgnc_wtEntrez")
}
\description{
A dataset containing all HGNC symbols in first column,
then entrez in second column.
}
\keyword{datasets}
