% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate.celltype.enrichment.probabilities.wtLimma.r
\name{calculate.celltype.enrichment.probabilities.wtLimma}
\alias{calculate.celltype.enrichment.probabilities.wtLimma}
\title{Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files}
\usage{
calculate.celltype.enrichment.probabilities.wtLimma(
  magmaAdjZ,
  ctd,
  thresh = 1e-04,
  sctSpecies = "mouse"
)
}
\arguments{
\item{magmaAdjZ}{Output from adjust.zstat.in.genesOut()}

\item{ctd}{Cell type data structure}

\item{thresh}{A threshold on low specificity values (used to drop genes)}

\item{sctSpecies}{Either 'human' or 'mouse'}
}
\description{
Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files
}
\examples{
magmaGenesOut = adjust.zstat.in.genesOut(ctd,magma_file=
"/Users/natske/GWAS_Summary_Statistics/MAGMA_Files/20016.assoc.tsv.10UP.1.5DOWN/20016.assoc.tsv.10UP.1.5DOWN.genes.out",
sctSpecies="mouse")

}
