% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate.celltype.enrichment.probabilities.wtLimma.r
\name{calculate.celltype.enrichment.probabilities.wtLimma}
\alias{calculate.celltype.enrichment.probabilities.wtLimma}
\title{Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files}
\usage{
calculate.celltype.enrichment.probabilities.wtLimma(
  magmaAdjZ,
  ctd,
  thresh = 1e-04,
  sctSpecies = "mouse",
  annotLevel = 1,
  celltypes = NULL,
  return_all = F
)
}
\arguments{
\item{magmaAdjZ}{Output from adjust.zstat.in.genesOut()}

\item{ctd}{CellTypeData object}

\item{thresh}{A threshold on low specificity values (used to drop genes)}

\item{sctSpecies}{Either 'human' or 'mouse'}

\item{annotLevel}{CellTypeData level to test
#'}
}
\description{
Calculate celltype enrichments without MAGMA using adjusted MAGMA Z-statistic from .genes.out files
}
\examples{
library(MAGMA.Celltyping)
library(ewceData)
ctd <- ewceData::ctd()

 # The package stores an example genesOut file, so save this to a tempfile
myGenesOut = tempfile()
data.table::fwrite(MAGMA.Celltyping::genesOut,sep="\t",file=myGenesOut)
magmaAdjZ = adjust.zstat.in.genesOut(ctd, magma_GenesOut_file=myGenesOut,sctSpecies="mouse")
ps <- calculate.celltype.enrichment.probabilities.wtLimma(magmaAdjZ=magmaAdjZ, ctd=ctd)

}
