% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_celltype_associations.r
\name{calculate_celltype_associations}
\alias{calculate_celltype_associations}
\title{Calculate celltype associations using MAGMA}
\usage{
calculate_celltype_associations(ctd, gwas_sumstats_path,
  analysis_name = "MainRun", upstream_kb = 10, downstream_kb = 1.5,
  genome_ref_path, specificity_species = "mouse", genesOutCOND = NA)
}
\arguments{
\item{ctd}{Cell type data strucutre containing $quantiles}

\item{gwas_sumstats_path}{Filepath of the summary statistics file}

\item{analysis_name}{Used in filenames which area created}

\item{upstream_kb}{How many kb upstream of the gene should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene should SNPs be included?}

\item{genome_ref_path}{Path to the folder containing the 1000 genomes .bed files (which can be downloaded from https://ctg.cncr.nl/software/MAGMA/ref_data/g1000_eur.zip)}

\item{specificity_species}{Species name relevant to the cell type data, i.e. "mouse" or "human"}

\item{genesOutCOND}{[Optional] Path to a genes.out file to condition on. Used if you want to condition on a different GWAS.}
}
\value{
Filepath for the genes.out file
}
\description{
Assumes that you have already run map.snps.to.genes()
}
\examples{
ctAssocs = calculate_celltype_associations(ctd,gwas_sumstats_path)

}
