% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_celltype_associations.r
\name{calculate_celltype_associations}
\alias{calculate_celltype_associations}
\title{Calculate celltype associations using MAGMA}
\usage{
calculate_celltype_associations(
  ctd,
  ctd_levels = seq_len(length(ctd)),
  ctd_species = infer_ctd_species(ctd),
  gwas_sumstats_path = NULL,
  magma_dir = NULL,
  analysis_name = "MainRun",
  prepare_ctd = TRUE,
  upstream_kb = 35,
  downstream_kb = 10,
  genesOutCOND = NA,
  EnrichmentMode = "Linear",
  force_new = FALSE,
  version = NULL,
  verbose = TRUE
)
}
\arguments{
\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_levels}{Which levels of \code{ctd} to
iterate the enrichment analysis over.}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}

\item{gwas_sumstats_path}{File path of the summary statistics file.}

\item{magma_dir}{Path to folder containing the
pre-computed MAGMA GWAS files (\emph{.gsa.raw}and \emph{.gsa.out}).}

\item{analysis_name}{Used in file names which area created.}

\item{prepare_ctd}{Whether to run
\link[MAGMA.Celltyping]{prepare_quantile_groups} on the \code{ctd} first.}

\item{upstream_kb}{How many kb upstream of the gene
should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene
should SNPs be included?}

\item{genesOutCOND}{[Optional] Path to a genes.out file to condition on.
Used if you want to condition on a different GWAS.}

\item{EnrichmentMode}{[Optional] Should either 'Linear' or 'Top 10\%' mode
be used for testing enrichment?}

\item{force_new}{[Optional] Force new MAGMA analyses even if the
pre-existing results files are detected.}

\item{version}{MAGMA version to use.}

\item{verbose}{Print messages.}
}
\value{
File path for the genes.out file
}
\description{
Assumes that you have already run
\link[MAGMA.Celltyping]{map_snps_to_genes}.
}
\examples{
#### Prepare cell-type data ####
ctd <- ewceData::ctd()

#### Prepare GWAS MAGMA data ####
magma_dir <- MAGMA.Celltyping::import_magma_files(ids = "ieu-a-298")
    
#### Run pipeline ####
ctAssocs <- calculate_celltype_associations(
    ctd = ctd,
    ctd_levels = 1,
    magma_dir = magma_dir,
    ctd_species = "mouse")
}
\keyword{main_function}
