% Generated by roxygen2: do not edit by hand
% Please edit documentation in
%   R/calculate_celltype_enrichment_probabilities_wtLimma.r
\name{calculate_celltype_enrichment_probabilities_wtLimma}
\alias{calculate_celltype_enrichment_probabilities_wtLimma}
\title{Calculate celltype enrichment}
\usage{
calculate_celltype_enrichment_probabilities_wtLimma(...)
}
\arguments{
\item{magmaAdjZ}{Output from
\link[MAGMA.Celltyping]{adjust_zstat_in_genesOut}.}

\item{thresh}{A threshold on low specificity values (used to drop genes).}
}
\description{
Calculate celltype enrichments without MAGMA
using adjusted MAGMA Z-statistic from .genes.out files.
}
\examples{
ctd <- ewceData::ctd()

# The package stores an example genesOut file, so save this to a tempfile
myGenesOut <- tempfile()
data.table::fwrite(
    x = MAGMA.Celltyping::genesOut,
    sep = "\t",
    file = myGenesOut
)
magmaAdjZ <- MAGMA.Celltyping::adjust_zstat_in_genesOut(
    ctd = ctd,
    magma_GenesOut_file = myGenesOut
)
ps <- calculate_celltype_enrichment_probabilities_wtLimma(
    magmaAdjZ = magmaAdjZ,
    ctd = ctd
)
}
