% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_conditional_celltype_associations.r
\name{calculate_conditional_celltype_associations}
\alias{calculate_conditional_celltype_associations}
\title{Calculate conditional celltype associations using MAGMA}
\usage{
calculate_conditional_celltype_associations(
  ctd,
  gwas_sumstats_path,
  analysis_name = "MainRun",
  upstream_kb = 10,
  downstream_kb = 1.5,
  genome_ref_path,
  controlledAnnotLevel = 1,
  specificity_species = "mouse",
  controlTopNcells = NA,
  controlledCTs = NA,
  EnrichmentMode = "Linear"
)
}
\arguments{
\item{ctd}{Cell type data strucutre containing $specificity_quantiles}

\item{gwas_sumstats_path}{Filepath of the summary statistics file}

\item{analysis_name}{Used in filenames which area created}

\item{upstream_kb}{How many kb upstream of the gene should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the
gene should SNPs be included?}

\item{genome_ref_path}{Path to the folder containing the 1000 genomes
.bed files, downladed using \link[MAGMA.Celltyping]{get_genome_ref}.}

\item{controlledAnnotLevel}{Which annotation level should be controlled for.}

\item{specificity_species}{Species name relevant to the cell type
data (e.g. 'mouse' or 'human').}

\item{controlTopNcells}{How many of the most significant cell types at
that annotation level should be controlled for?}

\item{controlledCTs}{Array of the celltype to be controlled for,
e.g. \code{c('Interneuron type 16','Medium Spiny Neuron')}.}

\item{EnrichmentMode}{[Optional] Either 'Linear' or 'Top 10\%'.
(\emph{DEFAULT: "Linear"}).}
}
\value{
Filepath for the genes.out file
}
\description{
Assumes that you have already run map_snps_to_genes()
}
\examples{
ctAssocs <- calculate_celltype_associations(ctd, gwas_sumstats_path)
}
