% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/celltype_associations_pipeline.r
\name{celltype_associations_pipeline}
\alias{celltype_associations_pipeline}
\title{Calculate celltype associations using MAGMA}
\usage{
celltype_associations_pipeline(
  ctd,
  ctd_name,
  magma_dirs,
  genome_ref_path,
  specificity_species = "mouse",
  run_linear = T,
  run_top10 = T,
  run_condition = F,
  upstream_kb = 35,
  downstream_kb = 10,
  suffix_linear = "linear",
  suffix_top10 = "top10",
  suffix_condition = "condition",
  controlTopNcells = 1,
  force_new = F,
  save_dir = tempdir()
)
}
\arguments{
\item{ctd}{Cell type data structure containing $specificity_quantiles}

\item{ctd_name}{Used in file names}

\item{magma_dirs}{Names of folders containing the pre-computed MAGMA GWAS files. 
\bold{NOTE}: Files within these folders must have the same naming scheme as the folders themselves.}

\item{genome_ref_path}{Path to the folder containing the 1000 genomes .bed files (which can be downloaded 
from https://ctg.cncr.nl/software/MAGMA/ref_data/g1000_eur.zip)}

\item{specificity_species}{Species name relevant to the cell type data, i.e. "mouse" or "human"}

\item{run_linear}{Run \code{calculate_celltype_associations} in linear mode.}

\item{run_top10}{Run \code{calculate_celltype_associations} in top 10% mode.}

\item{run_condition}{Run \code{calculate_celltype_associations} in conditional mode.}

\item{upstream_kb}{How many kb upstream of the gene should SNPs be included?}

\item{downstream_kb}{How many kb downstream of the gene should SNPs be included?}

\item{suffix_linear}{This will be added to the linear results file name.}

\item{suffix_top10}{This will be added to the top 10% results file name.}

\item{suffix_condition}{This will be added to the conditional results file name.}

\item{controlTopNcells}{How many of the most significant cell types at that annotation level should be controlled for?}

\item{force_new}{[Optional] Force new MAGMA analyses even if the pre-existing results files are detected.}

\item{save_dir}{Folder to save results in (\code{save_dir=NULL} to not save any results).}
}
\value{
A list containing the results of each selected celltype associations analysis.
}
\description{
Has the option of running multiple analyses with a single function.  
Assumes that you already have all MAGMA GWAS files precomputed.
Precomputed MAGMA GWAS files can be downlaoded via the \code{import_magma_files} fucnction.
}
\examples{
\dontrun{
library(MAGMA.Celltyping)
local_files <- import_magma_files(download_dir=".")
magma_dirs <- unique(dirname(local_files))
res <- celltype_associations_pipeline(ctd=ewceData::ctd(), ctd_name="Zeisel2018", magma_dirs=magma_dirs, genome_ref_path="~/Downloads/g1000_eur/g1000_eur")
}
}
