% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_controlledCTs.R
\name{check_controlledCTs}
\alias{check_controlledCTs}
\title{Check controlled cell type names}
\usage{
check_controlledCTs(ctd, controlledCTs, controlledAnnotLevel)
}
\description{
Check the cell types in 'controlledCT' exist
at the relevant annotation level.
}
