% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare.trait.enrichment.r
\name{compare.trait.enrichments}
\alias{compare.trait.enrichments}
\title{Compare Trait Enrichments}
\usage{
compare.trait.enrichments(magmaPath1 = NA, magmaPath2 = NA,
  magma1 = NA, magma2 = NA, annotLevel, ctd)
}
\arguments{
\item{magmaPath1}{A magma results file}

\item{magmaPath2}{A magma results file}

\item{magma1}{A magma results data structure}

\item{magma2}{A magma results data structure}

\item{annotLevel}{Annotation level of the CTD / results file to use}

\item{ctd}{The celltypedata structure}
}
\value{
Results
}
\description{
Test to see which cell types are more enriched in magmaPath2 than magmaPath1
}
\examples{
1==1

}
