% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_genesets.R
\name{create_genesets}
\alias{create_genesets}
\title{Create gene sets}
\usage{
create_genesets(
  res_input,
  n_genes = 100,
  spec_deciles = 10,
  human.symbol = TRUE
)
}
\arguments{
\item{res_input}{Enrichment results output from 
\link[MAGMA.Celltyping]{calculate_celltype_enrichment_limma}.}

\item{n_genes}{Max number of genes to return.}

\item{spec_deciles}{Which specificity deciles to use
from the CellTypeDataset.}

\item{human.symbol}{Whether to use human gene symbols.}
}
\description{
Support function for \link[MAGMA.Celltyping]{get_driver_genes}.
}
\keyword{internal}
