% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_top10percent_genesets_file.R
\name{create_top10percent_genesets_file}
\alias{create_top10percent_genesets_file}
\title{Create gene covar file}
\source{
\code{
#### Prepare cell-type data ####
ctd <- ewceData::ctd()

#### Prepare GWAS MAGMA data ####
myGenesOut <- MAGMA.Celltyping::import_magma_files(ids = "ieu-a-298",
                                                  file_types = ".genes.out",
                                                  return_dir = FALSE)
ctd <- MAGMA.Celltyping::prepare_quantile_groups(ctd = ctd)
geneSetsFilePath <- MAGMA.Celltyping:::create_top10percent_genesets_file(
    genesOutFile = myGenesOut,
    ctd = ctd,
    annotLevel = 1,
    ## Mapped to human orths by prepare_quantile_groups previously
    ctd_species = "human"
)
}
}
\usage{
create_top10percent_genesets_file(
  genesOutFile,
  ctd,
  annotLevel,
  ctd_species,
  verbose = TRUE
)
}
\arguments{
\item{genesOutFile}{The output of the second call to MAGMA
(performed in the map_snps_to_genes function).}

\item{ctd}{Cell type data structure containing
\code{specificity_quantiles}.}

\item{ctd_species}{Species name relevant to the CellTypeDataset (\code{ctd}).
See \link[EWCE]{list_species} for all available species. 
If \code{ctd_species=NULL} (default),
 the \code{ctd} species will automatically 
 be inferred using \link[orthogene]{infer_species}.}

\item{verbose}{Print messages.}
}
\value{
Filepath for the gene covar file
}
\description{
The gene covar file is the input to MAGMA for the celltype
association analysis. This code was functonalised
because it is called by both baseline and conditional analysis.
}
\keyword{internal}
