% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_top10percent_genesets_file.R
\name{create_top10percent_genesets_file}
\alias{create_top10percent_genesets_file}
\title{Create gene covar file}
\usage{
create_top10percent_genesets_file(genesOutFile, ctd, annotLevel,
  specificity_species)
}
\arguments{
\item{genesOutFile}{The output of the second call to MAGMA (performed in the map.snps.to.genes function)}

\item{ctd}{Cell type data structure. Must contain quantiles.}

\item{annotLevel}{Annot level for which the gene covar file should be constructed}

\item{specificity_species}{Species name relevant to the cell type data, i.e. "mouse" or "human"}
}
\value{
Filepath for the gene covar file
}
\description{
The gene covar file is the input to MAGMA for the celltype association analysis. This code was functonalised because it is called by both baseline and conditional analysis.
}
\examples{
library(MAGMA.Celltyping)
genesOutFile="/Users/natske/Naomi_Wray_Conditional/Results/Conditional/scz_adj_all_withaut_gsmr_bxy_aric_reference.raw"
load("/Users/natske/Box Sync/Patrick & Julien/CellTypeData_allKImouse_wtHypo_MergedStriatal_1to1only.rda")
ctd = prepare.quantile.groups(ctd,specificity_species="mouse",numberOfBins=40)
geneSetsFilePath = create_top10percent_genesets_file(genesOutFile="",ctd=ctd,annotLevel=1,specificity_species="mouse")

}
