% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.r
\docType{data}
\name{ctd_allKI}
\alias{ctd_allKI}
\title{Celltype data (allKI)}
\format{An object of class \code{list} of length 2.}
\source{
# http://www.hjerling-leffler-lab.org/data/scz_singlecell/
usethis::use_data(ctd_allKI,overwrite=TRUE)
}
\usage{
ctd_allKI
}
\description{
CTD file with cortex, hippocampus, hypothalamus and midbrain
}
\keyword{datasets}
