% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompress.R
\name{decompress}
\alias{decompress}
\title{Decompress a file}
\usage{
decompress(
  path_formatted,
  remove = FALSE,
  overwrite = FALSE,
  storage_dir = tempdir(),
  verbose = TRUE
)
}
\arguments{
\item{path_formatted}{Path to compressed file.}

\item{remove}{If \code{\link[base:logical]{TRUE}}, the input file is removed afterward,
   otherwise not.}

\item{overwrite}{If \code{\link[base:logical]{TRUE}} and the output file already exists,
   the file is silently overwritten, otherwise an exception is
   thrown (unless \code{skip} is \code{\link[base:logical]{TRUE}}).}

\item{storage_dir}{Folder in which to store the GWAS summary stats.}

\item{verbose}{Print messages.}
}
\value{
Path to decompressed file.
}
\description{
Decompress a gzipped (.gz) or bgzipped (.bgz) file.
}
\keyword{internal}
