% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_ctd_genes.R
\name{filter_ctd_genes}
\alias{filter_ctd_genes}
\title{Filter genes in a CellTypeDataset}
\usage{
filter_ctd_genes(ctd, gene_subset)
}
\arguments{
\item{ctd}{CellTypeDataset}

\item{gene_subset}{Genes to subset to.}
}
\description{
Filter genes in a CellTypeDataset
}
\examples{
library(MAGMA.Celltyping)
library(ewceData)
set.seed(1)
ctd <- ewceData::ctd()
gene_subset <- sample(rownames(ctd[[1]]$mean_exp), 1000)
ctd_subset <- filter_ctd_genes(ctd = ctd, gene_subset = gene_subset)

}
