% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_sumstats_for_magma_macOnly.r
\name{format_sumstats_for_magma_macOnly}
\alias{format_sumstats_for_magma_macOnly}
\title{Check that sumstats has correct columns and that they are in the correct order for MAGMA and LDSC}
\usage{
format_sumstats_for_magma_macOnly(path)
}
\value{
col_headers The new column headers for the sumstats file
}
\description{
Check that sumstats has correct columns and that they are in the correct order for MAGMA and LDSC
}
\examples{
format_sumstats_for_magma_macOnly(path)

}
