% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gather_results.R
\name{gather_results}
\alias{gather_results}
\title{Gather enrichment results from \code{celltype_associations_pipeline}}
\usage{
gather_results(
  MAGMA_results,
  level = 2,
  dataset_name = NULL,
  species = "mouse",
  filetype = "ctAssocMerged",
  fdr_thresh = NULL,
  save_dir = NULL
)
}
\description{
Gather enrichment results from \code{celltype_associations_pipeline}
}
\examples{
\dontrun{
library(MAGMA.Celltyping)
local_files <- import_magma_files(download_dir=".")
#' magma_dirs <- unique(dirname(local_files))
res <- celltype_associations_pipeline(ctd=ewceData::ctd(), 
                                      ctd_name="Zeisel2018", 
                                      magma_dirs=magma_dirs, 
                                      genome_ref_path="~/Downloads/g1000_eur/g1000_eur") 
merged_res <- gather_results(res)
}
}
