% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ctd_dendro.r
\name{get_ctd_dendro}
\alias{get_ctd_dendro}
\title{Get CTD dendro}
\source{
\code{
MAGMA.Celltyping:::get_ctd_dendro(ctd = ctd, annotLevel = 1)
}
}
\usage{
get_ctd_dendro(ctd, annotLevel)
}
\arguments{
\item{ctd}{Cell type data strucutre containing \code{specificity_quantiles}.}

\item{annotLevel}{Annot level for which the gene covar file should
be constructed.}
}
\value{
List containing ddata and ordered_cells
}
\description{
Get all dendrogram features used for plotting a dendrogram of cell types.
}
\keyword{internal}
