% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{get data via \pkg{piggyback}}
\usage{
get_data(
  fname,
  repo = "neurogenomics/MAGMA_Celltyping",
  storage_dir = tempdir(),
  overwrite = FALSE,
  check = FALSE
)
}
\description{
get data via \pkg{piggyback}
}
\keyword{internal}
