% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get remote data}
\usage{
get_data(
  fname,
  repo = "neurogenomics/MAGMA_Celltyping",
  storage_dir = tools::R_user_dir(package = "MAGMA.Celltyping", which = "cache"),
  overwrite = TRUE,
  check = FALSE
)
}
\arguments{
\item{repo}{Repository name in format "owner/repo". Defaults to \code{guess_repo()}.}

\item{overwrite}{Should any local files of the same name be overwritten?
default \code{TRUE}.}
}
\description{
Download remotely stored data via \link[piggyback]{pb_download}.
}
\keyword{internal}
