% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_driver_genes.R
\name{get_driver_genes}
\alias{get_driver_genes}
\title{Get genes driving significant \pkg{MAGMA_celltyping} results}
\usage{
get_driver_genes(
  ctd,
  sctSpecies = "mouse",
  magma_res,
  fdr_thresh = 0.05,
  GenesOut_dir = "./",
  n_genes = 100,
  spec_deciles = 10
)
}
\arguments{
\item{ctd}{CellTypeData object}

\item{sctSpecies}{Either 'human' or 'mouse'}

\item{magma_res}{Merged results from \code{MAGMA.Celltyping::gather_results}.}

\item{fdr_thresh}{FDR threshold for \code{magma_res}.}

\item{GenesOut_dir}{Folder to search for \emph{.genes.out} files implicated in \code{magma_res}.}

\item{n_genes}{Max number of drive genes to return per cell-type enrichment.}

\item{spec_deciles}{Which \emph{specificity_proportion} deciles to include when calculating driver genes. 
(10 = most specific).}
}
\description{
Get genes driving significant \pkg{MAGMA_celltyping} results
}
