% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_genome_ref_piggyback.R
\name{get_genome_ref_piggyback}
\alias{get_genome_ref_piggyback}
\title{Download genome ref: piggyback}
\source{
\code{
genome_ref_path <- MAGMA.Celltyping::get_genome_ref(method = "magma")

piggyback::pb_upload(file = genome_ref_path,
                     repo = "neurogenomics/MAGMA_Celltyping",
                     overwrite = TRUE)
}
}
\usage{
get_genome_ref_piggyback(genome_ref_dir, verbose = TRUE)
}
\arguments{
\item{genome_ref_dir}{Folder to download reference to.}

\item{verbose}{Print messages.}
}
\description{
Download and decompress the genome reference needed to run
\link[MAGMA.Celltyping]{map_snps_to_genes},
If the file already exists, it will simply return the path.
}
\keyword{internal}
