% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/import_magma_files.r
\name{import_magma_files}
\alias{import_magma_files}
\title{Import remote MAGMA files}
\usage{
import_magma_files(
  download_dir = tempdir(),
  file_types = c(".genes.raw", ".genes.out"),
  overwrite = F,
  nThread = parallel::detectCores() - 2
)
}
\arguments{
\item{download_dir}{Parent folder where you want to save the MAGMA files.}

\item{overwrite}{If the files have already been downloaded in the specified directory, 
these downloads will be skipped. Set \code{overwrite=T} to force them to be re-downloaded.}

\item{nThread}{Number of threads to parallelize downloading.}
}
\value{
Paths to where the MAGMA files have been downloaded locally.
}
\description{
Import preprocessed MAGMA GWAS files from remote
 \href{https://github.com/neurogenomics/MAGMA_Files}{GitHub repository}.
Each MAGMA file will be put into its own folder of the same name as the file. 
\bold{WARNING}:
}
\examples{
local_files <- import_magma_files(download_dir=".")
}
