% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/install_magma.R
\name{install_magma}
\alias{install_magma}
\title{Install the MAGMA command line tool}
\source{
[MAGMA website](https://ctg.cncr.nl/software/magma)

[MAGMA.celltyping documentation](https://github.com/NathanSkene/MAGMA_Celltyping)
}
\usage{
install_magma(dest_dir = "~")
}
\description{
Install the MAGMA command line tool
}
