% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load.magma.results.file.r
\name{load.magma.results.file}
\alias{load.magma.results.file}
\title{Load a MAGMA .gcov.out file}
\usage{
load.magma.results.file(
  path,
  annotLevel,
  ctd,
  genesOutCOND = NA,
  EnrichmentMode = "Linear",
  ControlForCT = "BASELINE"
)
}
\arguments{
\item{path}{Path to a .gcov.out file (if EnrichmentMode=='Linear') or .sets.out (if EnrichmentMode=='Top 10\%')}

\item{annotLevel}{Which annotation level does this .gcov.out file relate to?}

\item{ctd}{Cell type data strucutre containing $specificity_quantiles}

\item{genesOutCOND}{[Optional] If the analysis controlled for another GWAS, then this is included as a column, otherwise the column is NA}

\item{EnrichmentMode}{[Optional] Either 'Linear' or 'Top 10\%'. Default assumes Linear.}

\item{ControlForCT}{[Optional] May be an internal argument. I'd suggest ignoring or take a look at the code to figure it out.}
}
\value{
Contents of the .gcov.out file
}
\description{
Convenience function which just does a little formatting to make it easier to use
}
\examples{
library(MAGMA.Celltyping)
ctd <- ewceData::ctd()  
res = load.magma.results.file(path="Raw/adhd_formtcojo.txt.NEW.10UP.1DOWN.gov.out",
                              annotLevel=1,ctd=ctd) 

}
