% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/load_magma_results_file.r
\name{load_magma_results_file}
\alias{load_magma_results_file}
\title{Load a MAGMA .gcov.out file}
\usage{
load_magma_results_file(
  path,
  annotLevel,
  ctd,
  genesOutCOND = NA,
  EnrichmentMode = "Linear",
  ControlForCT = "BASELINE"
)
}
\arguments{
\item{path}{Path to a .gcov.out file (if EnrichmentMode=='Linear')
or .sets.out (if EnrichmentMode=='Top 10\%').}

\item{genesOutCOND}{[Optional] If the analysis controlled for another GWAS, 
then this is included as a column. Otherwise the column is \code{NA}.}

\item{EnrichmentMode}{[Optional] Either 'Linear' or 'Top 10\%' 
(Default: \code{'Linear'}).}

\item{ControlForCT}{[Optional] May be an internal argument. 
We suggest ignoring or take a look at the code to figure it out.}
}
\value{
Contents of the .gcov.out file.
}
\description{
Convenience function which just does a little formatting to make 
it easier to use.
}
\examples{
\dontrun{
ctd <- ewceData::ctd()
res <- MAGMA.Celltyping::load_magma_results_file(
    path = "Raw/adhd_formtcojo.txt.NEW.10UP.1DOWN.gcov.out",
    annotLevel = 1,
    ctd = ctd
)
} 
}
