% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_install.R
\name{magma_install}
\alias{magma_install}
\title{Install the MAGMA command line tool}
\source{
\href{https://ctg.cncr.nl/software/magma}{MAGMA website}

\href{https://github.com/neurogenomics/MAGMA_Celltyping}{
MAGMA.celltyping documentation}
}
\usage{
magma_install(
  dest_dir = find_install_dir(),
  desired_version = "latest",
  upgrade = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dest_dir}{Folder in which to install MAGMA.}

\item{desired_version}{Desired version of MAGMA.}

\item{upgrade}{If MAGMA is already installed and is not the latest version, 
should it be upgraded to the latest version? (Default: \code{FALSE}).}

\item{verbose}{Print messages.}
}
\value{
Path to MAGMA executable.
}
\description{
Checks whether MAGMA is currently installed, and if not, 
tries to automatically install it.
}
\examples{
magma <- MAGMA.Celltyping::magma_install()
}
