% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_installed_version.R
\name{magma_installed_version}
\alias{magma_installed_version}
\title{MAGMA installed version}
\usage{
magma_installed_version(
  magma_x = magma_executable(),
  as_package_version = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{magma_x}{MAGMA executable.}

\item{as_package_version}{Return as the object class 
\link[base]{package_version}. 
NOTE, this will convert any character suffixes into numbers:
e.g. \code{"v1.06a"} --> \code{"v1.06.1"}, or 
\code{"v1.06b"} --> \code{"v1.06.2"}}

\item{verbose}{Print messages.}
}
\description{
List which version(s) of MAGMA is currently installed.
}
\keyword{internal}
