% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/magma_tileplot.r
\name{magma_tileplot}
\alias{magma_tileplot}
\title{MAGMA tileplot}
\usage{
magma_tileplot(
  ctd,
  results,
  height = 13,
  width = 4,
  annotLevel = 1,
  fileTag = "",
  output_path
)
}
\arguments{
\item{ctd}{Cell type data structure containing "specificity_quantiles".}

\item{results}{Output from either 
\link[MAGMA.Celltyping]{calculate_celltype_associations} or
\link[MAGMA.Celltyping]{calculate_conditional_celltype_associations}.}

\item{height}{Height of the output tileplot.}

\item{width}{Width of the output tileplot.}

\item{annotLevel}{Annotation level to plot the results for.}

\item{fileTag}{String appended to the names of the saved PDFs,
i.e. the name of the celltype data file used.}

\item{output_path}{Location where the results should be plotted.}
}
\value{
List of two ggplot objects.
}
\description{
Used after merging results from multiple GWAS studies.
}
\examples{
 # 
}
