% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_specificity_to_entrez.r
\name{map_specificity_to_entrez}
\alias{map_specificity_to_entrez}
\title{Map specificity to Entrez gene IDs}
\usage{
map_specificity_to_entrez(
  ctd,
  annotLevel,
  sctSpecies,
  return_ctd = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{ctd}{Cell type data structure containing $specificity_quantiles.}

\item{sctSpecies}{Species name relevant to the cell type data.
See \link[EWCE]{list_species} for all available species.}

\item{return_ctd}{Return the actual CellTypeDataset,
rather than path to where it is saved.}

\item{verbose}{Print messages.}
}
\value{
Matrix in which the first column is 'entrez'
and then the specificity decile for each cell type
}
\description{
Convenience function used in \link[MAGMA.Celltyping]{create_gene_covar_file}.
}
\keyword{internal}
