% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/map_specificity_to_entrez.r
\name{map_specificity_to_entrez}
\alias{map_specificity_to_entrez}
\title{Map specificity to entrez}
\usage{
map_specificity_to_entrez(ctd, annotLevel, specificity_species, return_ctd = F)
}
\arguments{
\item{ctd}{Cell type data structure. Must contain quantiles.}

\item{annotLevel}{Annot level for which the gene covar file should be constructed}

\item{specificity_species}{Species name relevant to the cell type data, i.e. "mouse" or "human"}
}
\value{
Matrix in which the first column is 'entrez' and then the specificity decile for each cell type
}
\description{
Convenience function used in 'create_gene_covar_file()'
}
\examples{
genesCovarFilePath = create_gene_covar_file(genesOutFile,ctd)

}
