% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_associations.r
\name{plot_celltype_associations}
\alias{plot_celltype_associations}
\title{Plot celltype associations calculated using MAGMA}
\usage{
plot_celltype_associations(ctAssocs, ctd, useSignificanceLine = TRUE,
  savePDF = TRUE, fileTag = "", plotDendro = TRUE, gwas_title = "",
  plotLegend = TRUE)
}
\arguments{
\item{ctAssocs}{Output from either calculate_celltype_associations() or calculate_conditional_celltype_associations()}

\item{ctd}{Cell type data strucutre containing $quantiles (required to calculate dendrogram)}

\item{useSignificanceLine}{TRUE or FALSE. Should their be a vertical line marking bonferroni signifiance?}

\item{savePDF}{TRUE or FALSE. Save figure to file or print to screen?}

\item{fileTag}{String apprended to the names of the saved PDFs, i.e. the name of the celltype data file used}
}
\description{
Can take input from either calculate_celltype_associations() or calculate_conditional_celltype_associations()
}
\examples{
ctAssocs = calculate_celltype_associations(ctd,gwas_sumstats_path)

}
