% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_heatmap.R
\name{results_heatmap}
\alias{results_heatmap}
\title{Plot heatmap of enrichment results}
\usage{
results_heatmap(
  merged_results,
  title = NULL,
  x_lab = NULL,
  fdr_thresh = 0.05,
  facet_formula = "EnrichmentMode ~ .",
  x_var = "Celltype",
  y_var = "GWAS",
  fill_var = "-log1p(FDR)",
  scales = "free_y",
  space = "fixed",
  show_plot = TRUE,
  height = 5,
  width = 7,
  dpi = 300,
  save_path = file.path(tempdir(), "MAGMA_Celltyping.heatmap.jpg")
)
}
\arguments{
\item{merged_results}{Enrichment results generated by 
\link[MAGMA.Celltyping]{celltype_associations_pipeline} and merged by 
\link[MAGMA.Celltyping]{merge_results}.}

\item{title}{Plot title.}

\item{x_lab}{Plot x-axis label.}

\item{fdr_thresh}{FDR filtering threshold.}

\item{facet_formula}{Facet formula for plot,
passed to \link[ggplot2]{facet_grid}.}

\item{x_var}{x-axis variable.}

\item{y_var}{y-axis variable.}

\item{fill_var}{Fill variable.}

\item{scales}{Are scales shared across all facets (the default,
\code{"fixed"}), or do they vary across rows (\code{"free_x"}),
columns (\code{"free_y"}), or both rows and columns (\code{"free"})?}

\item{space}{If \code{"fixed"}, the default, all panels have the same size.
If \code{"free_y"} their height will be proportional to the length of the
y scale; if \code{"free_x"} their width will be proportional to the
length of the x scale; or if \code{"free"} both height and width will
vary.  This setting has no effect unless the appropriate scales also vary.}

\item{show_plot}{Whether to print the plot.}

\item{height}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{width}{Plot size in \code{units} ("in", "cm", "mm", or "px").
If not supplied, uses the size of current graphics device.}

\item{dpi}{Plot resolution. Also accepts a string input: "retina" (320),
"print" (300), or "screen" (72). Applies only to raster output types.}

\item{save_path}{Path to save plot to.}
}
\value{
ggplot object.
}
\description{
Plot of GWAS x cell-type enrichment results from 
\link[MAGMA.Celltyping]{celltype_associations_pipeline}.
}
\examples{
MAGMA_results <- MAGMA.Celltyping::enrichment_results
merged_results <- MAGMA.Celltyping::merge_results(MAGMA_results)
heat <- MAGMA.Celltyping::results_heatmap(
    merged_results = merged_results,
    fdr_thresh = 1)  
}
