% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/results_heatmap.R
\name{results_heatmap}
\alias{results_heatmap}
\title{Plot enrichment results from \code{celltype_associations_pipeline}}
\usage{
results_heatmap(
  merged_results,
  title = NULL,
  x_lab = NULL,
  fdr_thresh = 0.05,
  facet_formula = "EnrichmentMode ~ .",
  x_var = "Celltype",
  y_var = "GWAS",
  fill_var = "-log1p(FDR)",
  scales = "free_y",
  space = "fixed",
  show_plot = T
)
}
\description{
Plot enrichment results from \code{celltype_associations_pipeline}
}
\examples{
\dontrun{
library(MAGMA.Celltyping)
local_files <- import_magma_files(download_dir=".")
#' magma_dirs <- unique(dirname(local_files))
res <- celltype_associations_pipeline(ctd=ewceData::ctd(), 
                                      ctd_name="Zeisel2018", 
                                      magma_dirs=magma_dirs, 
                                      genome_ref_path="~/Downloads/g1000_eur/g1000_eur") 
merged_results <- gather_results(res)
heat <- results_heatmap(merged_results)
}
}
