% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardise.sumstats.column.headers.r
\name{standardise.sumstats.column.headers}
\alias{standardise.sumstats.column.headers}
\title{Standardise the column headers in the Summary Statistics files}
\usage{
standardise.sumstats.column.headers(path)
}
\arguments{
\item{path}{Filepath for the summary statistics file}
}
\value{
The amended column headers (also the column headers will be written directly into the summary statistics file)
}
\description{
Use a reference data table of common column header names (stored in sumstatsColHeaders.rda) convert them to a standard set, i.e. chromosome --> CHR
}
\details{
This function does not check that all the required column headers are present

The amended header is written directly back into the file
}
\examples{
col_headers <- standardise.sumstats.column.headers("~/Downloads/202040.assoc.tsv")
}
